% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.plot.BinaryPlot.R
\name{BinaryPlot}
\alias{BinaryPlot}
\title{Plotting feature clusters following bi-class problem.}
\description{
The \link{BinaryPlot} implements a basic plot for bi-class problem.
}
\seealso{
\code{\link{GenericPlot}}

for more information see \code{plot} function from \code{GenericPlot} class.
}
\keyword{color}
\keyword{device}
\keyword{hplot}
\section{Super class}{
\code{\link[DDMCS:GenericPlot]{DDMCS::GenericPlot}} -> \code{BinaryPlot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BinaryPlot$new()}}
\item \href{#method-plot}{\code{BinaryPlot$plot()}}
\item \href{#method-clone}{\code{BinaryPlot$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
empty function used to initalize the object arguments in runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryPlot$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{BinaryPlot} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
plots feature-clustering data from a bi-class problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryPlot$plot(summary, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary}}{a \link{data.frame} comprising the elements to be plotted.}

\item{\code{...}}{further arguments passed down to \code{plot} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
