% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.ClassificationOutput.R
\docType{class}
\name{ClassificationOutput}
\alias{ClassificationOutput}
\title{<<tittle>>}
\description{
ClassificationOutput
}
\details{
<<details>
}
\seealso{
\code{\link{DDMCS}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClassificationOutput$new()}}
\item \href{#method-getMetrics}{\code{ClassificationOutput$getMetrics()}}
\item \href{#method-getPositiveClass}{\code{ClassificationOutput$getPositiveClass()}}
\item \href{#method-getModelInfo}{\code{ClassificationOutput$getModelInfo()}}
\item \href{#method-getPerformances}{\code{ClassificationOutput$getPerformances()}}
\item \href{#method-savePerformances}{\code{ClassificationOutput$savePerformances()}}
\item \href{#method-plotPerformances}{\code{ClassificationOutput$plotPerformances()}}
\item \href{#method-getPredictions}{\code{ClassificationOutput$getPredictions()}}
\item \href{#method-savePredictions}{\code{ClassificationOutput$savePredictions()}}
\item \href{#method-clone}{\code{ClassificationOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$new(voting.schemes, models)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{voting.schemes}}{<<description>>}

\item{\code{models}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetrics"></a>}}
\subsection{Method \code{getMetrics()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$getMetrics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPositiveClass"></a>}}
\subsection{Method \code{getPositiveClass()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$getPositiveClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getModelInfo"></a>}}
\subsection{Method \code{getModelInfo()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$getModelInfo(metrics = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metrics}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPerformances"></a>}}
\subsection{Method \code{getPerformances()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$getPerformances(
  test.set,
  measures,
  voting.names = NULL,
  metric.names = NULL,
  cutoff.values = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test.set}}{<<description>>}

\item{\code{measures}}{<<description>>}

\item{\code{voting.names}}{<<description>>}

\item{\code{metric.names}}{<<description>>}

\item{\code{cutoff.values}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-savePerformances"></a>}}
\subsection{Method \code{savePerformances()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$savePerformances(
  dir.path,
  test.set,
  measures,
  voting.names = NULL,
  metric.names = NULL,
  cutoff.values = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{<<description>>}

\item{\code{test.set}}{<<description>>}

\item{\code{measures}}{<<description>>}

\item{\code{voting.names}}{<<description>>}

\item{\code{metric.names}}{<<description>>}

\item{\code{cutoff.values}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPerformances"></a>}}
\subsection{Method \code{plotPerformances()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$plotPerformances(
  dir.path,
  test.set,
  measures,
  voting.names = NULL,
  metric.names = NULL,
  cutoff.values = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{<<description>>}

\item{\code{test.set}}{<<description>>}

\item{\code{measures}}{<<description>>}

\item{\code{voting.names}}{<<description>>}

\item{\code{metric.names}}{<<description>>}

\item{\code{cutoff.values}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPredictions"></a>}}
\subsection{Method \code{getPredictions()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$getPredictions(
  voting.names = NULL,
  metric.names = NULL,
  cutoff.values = NULL,
  type = NULL,
  target = NULL,
  filter = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{voting.names}}{<<description>>}

\item{\code{metric.names}}{<<description>>}

\item{\code{cutoff.values}}{<<description>>}

\item{\code{type}}{<<description>>}

\item{\code{target}}{<<description>>}

\item{\code{filter}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-savePredictions"></a>}}
\subsection{Method \code{savePredictions()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$savePredictions(
  dir.path,
  voting.names = NULL,
  metric.names = NULL,
  cutoff.values = NULL,
  type = NULL,
  target = NULL,
  filter = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{<<description>>}

\item{\code{voting.names}}{<<description>>}

\item{\code{metric.names}}{<<description>>}

\item{\code{cutoff.values}}{<<description>>}

\item{\code{type}}{<<description>>}

\item{\code{target}}{<<description>>}

\item{\code{filter}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassificationOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
