% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.ClassMajorityVoting.R
\name{ClassMajorityVoting}
\alias{ClassMajorityVoting}
\title{Implementation of Mayority Voting voting.}
\description{
Implementation of the parliamentary 'majority voting' procedure. The majority class value is defined as final class.
All class values have the same importance.
}
\seealso{
\code{\link{DDMCS}}, \code{\link{ClassMajorityVoting}}, \code{\link{ClassWeightedVoting}},
\code{\link{ProbAverageVoting}}, \code{\link{ProbAverageWeightedVoting}}, \code{\link{ProbBasedMethodology}}
}
\keyword{math}
\keyword{methods}
\keyword{models}
\section{Super class}{
\code{\link[DDMCS:SimpleVoting]{DDMCS::SimpleVoting}} -> \code{ClassMajorityVoting}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClassMajorityVoting$new()}}
\item \href{#method-getMajorityClass}{\code{ClassMajorityVoting$getMajorityClass()}}
\item \href{#method-getClassTie}{\code{ClassMajorityVoting$getClassTie()}}
\item \href{#method-execute}{\code{ClassMajorityVoting$execute()}}
\item \href{#method-clone}{\code{ClassMajorityVoting$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SimpleVoting" data-id="getCutoff">}\href{../../DDMCS/html/SimpleVoting.html#method-getCutoff}{\code{DDMCS::SimpleVoting$getCutoff()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SimpleVoting" data-id="getFinalPred">}\href{../../DDMCS/html/SimpleVoting.html#method-getFinalPred}{\code{DDMCS::SimpleVoting$getFinalPred()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMajorityVoting$new(cutoff = 0.5, class.tie = NULL, majority.class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cutoff}}{A \code{\link{character}} vector defining the minimum probability used to perform a
a positive classification. If is not defined, 0.5 will be used as default value.}

\item{\code{class.tie}}{A \code{\link{character}} used to define the target class value used when a tie is found.
If \code{\link{NULL}} positive class value will be assigned.}

\item{\code{majority.class}}{A \code{\link{character}} defining the value of the majority class.
If \code{\link{NULL}} will be used same value as trainning stage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMajorityClass"></a>}}
\if{latex}{\out{\hypertarget{method-getMajorityClass}{}}}
\subsection{Method \code{getMajorityClass()}}{
The function returns the value of the majority class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMajorityVoting$getMajorityClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link{character}} vector of length 1 with the name of the majority class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassTie"></a>}}
\if{latex}{\out{\hypertarget{method-getClassTie}{}}}
\subsection{Method \code{getClassTie()}}{
The function gets the class value assigned to solve ties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMajorityVoting$getClassTie()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link{character}} vector of length 1.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
The function implements the majority voting procedure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMajorityVoting$execute(predictions, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{A \code{\link{ClusterPredictions}} object containing all the predictions achieved for each cluster.}

\item{\code{verbose}}{A \code{\link{logical}} value to specify if more verbosity is needed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMajorityVoting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
