% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.ClassWeightedVoting.R
\docType{class}
\name{ClassWeightedVoting}
\alias{ClassWeightedVoting}
\title{Implementation Weighted Voting scheme.}
\description{
An new implementation of \code{\link{ClassMajorityVoting}} where each class value has different values (weights).
}
\seealso{
\code{\link{DDMCS}}, \code{\link{ClassMajorityVoting}}, \code{\link{ClassWeightedVoting}},
\code{\link{ProbAverageVoting}}, \code{\link{ProbAverageWeightedVoting}}, \code{\link{ProbBasedMethodology}}
}
\keyword{math}
\keyword{methods}
\keyword{models}
\section{Super class}{
\code{\link[DDMCS:SimpleVoting]{DDMCS::SimpleVoting}} -> \code{ClassWeightedVoting}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClassWeightedVoting$new()}}
\item \href{#method-getWeights}{\code{ClassWeightedVoting$getWeights()}}
\item \href{#method-setWeights}{\code{ClassWeightedVoting$setWeights()}}
\item \href{#method-execute}{\code{ClassWeightedVoting$execute()}}
\item \href{#method-clone}{\code{ClassWeightedVoting$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SimpleVoting" data-id="getCutoff">}\href{../../DDMCS/html/SimpleVoting.html#method-getCutoff}{\code{DDMCS::SimpleVoting$getCutoff()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SimpleVoting" data-id="getFinalPred">}\href{../../DDMCS/html/SimpleVoting.html#method-getFinalPred}{\code{DDMCS::SimpleVoting$getFinalPred()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassWeightedVoting$new(cutoff = 0.5, weights = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cutoff}}{A \code{\link{character}} vector defining the minimum probability used to perform a
a positive classification. If is not defined, 0.5 will be used as default value.}

\item{\code{weights}}{A \code{\link{numeric}} vector with the weigths of each cluster. If \code{\link{NULL}} performance
achieved during training will be used as default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWeights"></a>}}
\subsection{Method \code{getWeights()}}{
The funtion returns the weights used to perform the voting scheme.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassWeightedVoting$getWeights()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link{numeric}} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setWeights"></a>}}
\subsection{Method \code{setWeights()}}{
The funtion allows changing the value of the weights.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassWeightedVoting$setWeights(weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weights}}{A \code{\link{numeric}} vector containing the new weights.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
The function implements the cluster-weighted majority voting procedure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassWeightedVoting$execute(predictions, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{A \code{\link{ClusterPredictions}} object containing all the predictions achieved for each cluster.}

\item{\code{verbose}}{A \code{\link{logical}} value to specify if more verbosity is needed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassWeightedVoting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
