% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.CombinedMetrics.R
\docType{class}
\name{CombinedMetrics}
\alias{CombinedMetrics}
\title{<<tittle>>}
\description{
CombinedMetrics
}
\details{
<<details>
}
\seealso{
\code{\link{CombinedVoting}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CombinedMetrics$new()}}
\item \href{#method-getRequiredMetrics}{\code{CombinedMetrics$getRequiredMetrics()}}
\item \href{#method-getFinalPrediction}{\code{CombinedMetrics$getFinalPrediction()}}
\item \href{#method-clone}{\code{CombinedMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedMetrics$new(required.metrics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{required.metrics}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRequiredMetrics"></a>}}
\subsection{Method \code{getRequiredMetrics()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedMetrics$getRequiredMetrics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFinalPrediction"></a>}}
\subsection{Method \code{getFinalPrediction()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedMetrics$getFinalPrediction(
  raw.pred,
  prob.pred,
  positive.class,
  negative.class
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw.pred}}{<<description>>}

\item{\code{prob.pred}}{<<description>>}

\item{\code{positive.class}}{<<description>>}

\item{\code{negative.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
