% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.CombinedVoting.R
\docType{class}
\name{CombinedVoting}
\alias{CombinedVoting}
\title{<<tittle>>}
\description{
CombinedVoting
}
\details{
<<details>
}
\seealso{
\code{\link{VotingStrategy}}
}
\section{Super class}{
\code{\link[DDMCS:VotingStrategy]{DDMCS::VotingStrategy}} -> \code{CombinedVoting}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CombinedVoting$new()}}
\item \href{#method-getCombinedMetrics}{\code{CombinedVoting$getCombinedMetrics()}}
\item \href{#method-getMethodology}{\code{CombinedVoting$getMethodology()}}
\item \href{#method-getFinalPred}{\code{CombinedVoting$getFinalPred()}}
\item \href{#method-execute}{\code{CombinedVoting$execute()}}
\item \href{#method-clone}{\code{CombinedVoting$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="VotingStrategy" data-id="getMetrics">}\href{../../DDMCS/html/VotingStrategy.html#method-getMetrics}{\code{DDMCS::VotingStrategy$getMetrics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="VotingStrategy" data-id="getName">}\href{../../DDMCS/html/VotingStrategy.html#method-getName}{\code{DDMCS::VotingStrategy$getName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="VotingStrategy" data-id="getVotingSchemes">}\href{../../DDMCS/html/VotingStrategy.html#method-getVotingSchemes}{\code{DDMCS::VotingStrategy$getVotingSchemes()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedVoting$new(voting.schemes, combined.metrics, methodology, metrics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{voting.schemes}}{<<description>}

\item{\code{combined.metrics}}{<<description>}

\item{\code{methodology}}{<<description>}

\item{\code{metrics}}{<<description>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCombinedMetrics"></a>}}
\subsection{Method \code{getCombinedMetrics()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedVoting$getCombinedMetrics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMethodology"></a>}}
\subsection{Method \code{getMethodology()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedVoting$getMethodology()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFinalPred"></a>}}
\subsection{Method \code{getFinalPred()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedVoting$getFinalPred(type = NULL, target = NULL, filter = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{<<description>>}

\item{\code{target}}{<<description>>}

\item{\code{filter}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedVoting$execute(predictions, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{<<description>>}

\item{\code{verbose}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedVoting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
