% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.ConfMatrix.R
\docType{class}
\name{ConfMatrix}
\alias{ConfMatrix}
\title{<<tittle>>}
\description{
ConfMatrix
}
\details{
<<details>
}
\seealso{
\code{\link{DDMCS}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConfMatrix$new()}}
\item \href{#method-getConfusionMatrix}{\code{ConfMatrix$getConfusionMatrix()}}
\item \href{#method-getTP}{\code{ConfMatrix$getTP()}}
\item \href{#method-getTN}{\code{ConfMatrix$getTN()}}
\item \href{#method-getFN}{\code{ConfMatrix$getFN()}}
\item \href{#method-getFP}{\code{ConfMatrix$getFP()}}
\item \href{#method-clone}{\code{ConfMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfMatrix$new(confMatrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{confMatrix}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConfusionMatrix"></a>}}
\subsection{Method \code{getConfusionMatrix()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfMatrix$getConfusionMatrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTP"></a>}}
\subsection{Method \code{getTP()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfMatrix$getTP()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTN"></a>}}
\subsection{Method \code{getTN()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfMatrix$getTN()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFN"></a>}}
\subsection{Method \code{getFN()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfMatrix$getFN()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFP"></a>}}
\subsection{Method \code{getFP()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfMatrix$getFP()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
