% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2mcs.Options.R
\docType{class}
\name{d2mcs.Options}
\alias{d2mcs.Options}
\title{Object to handle the keys/attributes/options common on D2MCS}
\usage{
d2mcs.Options
}
\description{
This class provides the necessary methods to manage a list of
keys or options used along the MCS creation, both those provided by the default
library and those implemented by the user.
}
\section{Details}{

By default, the application initializes the object named \code{d2mcs.Options}
of type \code{D2MCSOptions} which is in charge of initializing the possible
options used in the D2MCS.

The default fields on \code{\link{d2mcs.Options}} are initialized, if needed,
as shown bellow:

\strong{[verbose]}

- \code{d2mcs.Options$set("verbose", <<status_verbose>>)}
}

\section{Log configuration}{


The d2mcs log is configured through the \code{configureLog} function.
This system manages both the place to display the messages and the priority
level of each message showing only the messages with a higher level than
indicated in the \emph{threshold} variable.

If you want to deactivate the d2mcs log, the \code{disableLog}
method in \code{\link{d2mcs.Options}} does this task.
}

\section{Methods}{

\itemize{
\item{\bold{get:}}{
obtains a specific option.
\itemize{
\item{\emph{Usage:}}{
\code{get(key)}
}
\item{\emph{Value:}}{
the value of the specific option.
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{key:}}{
(\emph{character}) the name of the option to obtain.
}
}
}
}
}

\item{\bold{add:}}{
adds a option to the list of options
\itemize{
\item{\emph{Usage:}}{
\code{add(key, value)}
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{key:}}{
(\emph{character}) the name of the new option.
}
\item{\strong{propertyName:}}{
(\emph{Object}) the value of the new option.
}
}
}
}
}

\item{\bold{set:}}{
modifies the value of the one option.
\itemize{
\item{\emph{Usage:}}{
\code{set(key, value)}
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{key:}}{
(\emph{character}) the name of the new option.
}
\item{\strong{propertyName:}}{
(\emph{Object}) the value of the new option.
}
}
}
}
}

\item{\bold{remove:}}{
removes a specific option.
\itemize{
\item{\emph{Usage:}}{
\code{remove(key)}
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{key:}}{
(\emph{character}) the name of the option to remove.
}
}
}
}
}

\item{\bold{getAll:}}{
gets the list of options.
\itemize{
\item{\emph{Usage:}}{
\code{getAll()}
}
\item{\emph{Value:}}{
Value of options.
}
}
}

\item{\bold{remove:}}{
resets the option list to the initial state.
\itemize{
\item{\emph{Usage:}}{
\code{reset()}
}
}
}

\item{\bold{isSpecificOption:}}{
checks for the existence of an specific option.
\itemize{
\item{\emph{Usage:}}{
\code{isSpecificProperty(key)}
}
\item{\emph{Value:}}{
A boolean results according to the existence of the specific option in the list of options
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{key:}}{
(\emph{character}) the key of the option to check.
}
}
}
}
}

\item{\bold{configureLog:}}{
Configures the D2MCS log.
\itemize{
\item{\emph{Usage:}}{
\code{configureLog(console = TRUE, threshold = "INFO", file = NULL)}
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{console:}}{
(\emph{boolean}) Shows the log on console or not.
}
\item{\strong{threshold:}}{
(\emph{character}) The logging threshold level. Messages with a lower
priority level will be discarded.
}
\item{\strong{file:}}{
(\emph{character}) The file to write messages to. If it is NULL,
the log in file will not be enabled.
}
}
}
}
}

\item{\bold{disableLog:}}{
Deactivates the D2MCS log.
\itemize{
\item{\emph{Usage:}}{
\code{disableLog()}
}
}
}

\item{\bold{getLogConfiguration:}}{
Print the D2MCS log configuration.
\itemize{
\item{\emph{Usage:}}{
\code{getLogConfiguration()}
}
}
}
}
}

\seealso{
\code{\link{d2mcs.log}}
}
