% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.DatasetLoader.R
\docType{class}
\name{DatasetLoader}
\alias{DatasetLoader}
\title{<<tittle>>}
\description{
DatasetLoader
}
\details{
<<details>
}
\seealso{
\code{\link{Dataset}}, \code{\link{HDDataset}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetLoader$new()}}
\item \href{#method-load}{\code{DatasetLoader$load()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetLoader$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\subsection{Method \code{load()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetLoader$load(
  filepath,
  header = TRUE,
  sep = ",",
  skip.lines = 0,
  target.class = NULL,
  positive.class = NULL,
  normalize.names = FALSE,
  string.as.factor = FALSE,
  ignore.columns = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filepath}}{<<description>>}

\item{\code{header}}{<<description>>}

\item{\code{sep}}{<<description>>}

\item{\code{skip.lines}}{<<description>>}

\item{\code{target.class}}{<<description>>}

\item{\code{positive.class}}{<<description>>}

\item{\code{normalize.names}}{<<description>>}

\item{\code{string.as.factor}}{<<description>>}

\item{\code{ignore.columns}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<descriptionReturn>>
}
}
}
