% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDMCS.R
\docType{class}
\name{DDMCS}
\alias{DDMCS}
\title{<<tittle>>}
\description{
DDMCS
}
\details{
<<details>
}
\seealso{
\code{\link{Dataset}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DDMCS$new()}}
\item \href{#method-train}{\code{DDMCS$train()}}
\item \href{#method-classify}{\code{DDMCS$classify()}}
\item \href{#method-getAvailableModels}{\code{DDMCS$getAvailableModels()}}
\item \href{#method-clone}{\code{DDMCS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDMCS$new(
  dir.path,
  num.cores = NULL,
  socket.type = "PSOCK",
  outfile = NULL,
  serialize = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{<<description>>}

\item{\code{num.cores}}{<<description>>}

\item{\code{socket.type}}{<<description>>}

\item{\code{outfile}}{<<description>>}

\item{\code{serialize}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDMCS$train(
  train.set,
  train.function,
  num.clusters = NULL,
  ex.classifiers = c(),
  ig.classifiers = c(),
  metrics = NULL,
  saveAllModels = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{train.set}}{<<description>>}

\item{\code{train.function}}{<<description>>}

\item{\code{num.clusters}}{<<description>>}

\item{\code{ex.classifiers}}{<<description>>}

\item{\code{ig.classifiers}}{<<description>>}

\item{\code{metrics}}{<<description>>}

\item{\code{saveAllModels}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-classify"></a>}}
\subsection{Method \code{classify()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDMCS$classify(train.output, subset, voting.types, positive.class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{train.output}}{<<description>>}

\item{\code{subset}}{<<description>>}

\item{\code{voting.types}}{<<description>>}

\item{\code{positive.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAvailableModels"></a>}}
\subsection{Method \code{getAvailableModels()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDMCS$getAvailableModels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDMCS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
