% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.utility.DefaultModelFit.R
\docType{class}
\name{DefaultModelFit}
\alias{DefaultModelFit}
\title{<<tittle>>}
\description{
DefaultModelFit
}
\details{
<<details>
}
\seealso{
\code{\link{GenericModelFit}}
}
\section{Super class}{
\code{\link[DDMCS:GenericModelFit]{DDMCS::GenericModelFit}} -> \code{DefaultModelFit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DefaultModelFit$new()}}
\item \href{#method-createFormula}{\code{DefaultModelFit$createFormula()}}
\item \href{#method-createRecipe}{\code{DefaultModelFit$createRecipe()}}
\item \href{#method-clone}{\code{DefaultModelFit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelFit$new(instances, class.name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instances}}{<<description>}

\item{\code{class.name}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createFormula"></a>}}
\subsection{Method \code{createFormula()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelFit$createFormula(simplify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simplify}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createRecipe"></a>}}
\subsection{Method \code{createRecipe()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelFit$createRecipe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
