% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.strategies.DependencyBasedStrategy.R
\docType{class}
\name{DependencyBasedStrategy}
\alias{DependencyBasedStrategy}
\title{Clustering strategy based on dependency between features}
\description{
Features are distributed according to their independence values.
This strategy is divided into two steps. The first phase focuses on forming
groups with those features most dependent on each other. This step also
identifies those that are independent from all the others in the group.
The second step is to try out different numbers of clusters until you find
the one you think is best. These clusters are formed by inserting in all the
independent characteristics identified previously and trying to distribute
the features of the groups formed in the previous step in separate clusters.
In this way, it seeks to ensure that the features are as independent as possible
from those found in the same cluster.
}
\details{
The strategy is suitable only for binary and real features. Other
features are automatically grouped into a specific cluster named as 'unclustered'.
This class requires the \code{\link{StrategyConfiguration}} type object implements the following methods:

- \code{getBinaryCutoff()}: The function is used to define the interval to consider the dependency between binary features.

- \code{getRealCutoff()}: The function allows defining the cutoff to consider the dependency between real features.

- \code{tiebreak(feature, clus.candidates, fea.dep.dist.clus, corpus, heuristic, class, class.name)}: The function solves the ties between two (or more) features.

- \code{qualityOfCluster(clusters, metrics)}: The function determines the quality of a cluster

- \code{isImprovingClustering(clusters.deltha)}: The function indicates if clustering is getting better as the number of them increases.

An example of implementation with the description of each parameter is the \code{\link{DependencyBasedStrategyConfiguration}} class.
}
\seealso{
\code{\link{ClusteringStrategy}}, \code{\link{StrategyConfiguration}}, \code{\link{DependencyBasedStrategyConfiguration}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[DDMCS:ClusteringStrategy]{DDMCS::ClusteringStrategy}} -> \code{DependencyBasedStrategy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DependencyBasedStrategy$new()}}
\item \href{#method-execute}{\code{DependencyBasedStrategy$execute()}}
\item \href{#method-getDistribution}{\code{DependencyBasedStrategy$getDistribution()}}
\item \href{#method-createTrain}{\code{DependencyBasedStrategy$createTrain()}}
\item \href{#method-plot}{\code{DependencyBasedStrategy$plot()}}
\item \href{#method-saveCSV}{\code{DependencyBasedStrategy$saveCSV()}}
\item \href{#method-clone}{\code{DependencyBasedStrategy$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getBestClusterDistribution">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getBestClusterDistribution}{\code{DDMCS::ClusteringStrategy$getBestClusterDistribution()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getConfiguration">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getConfiguration}{\code{DDMCS::ClusteringStrategy$getConfiguration()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getDescription">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getDescription}{\code{DDMCS::ClusteringStrategy$getDescription()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getHeuristic">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getHeuristic}{\code{DDMCS::ClusteringStrategy$getHeuristic()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getUnclustered">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getUnclustered}{\code{DDMCS::ClusteringStrategy$getUnclustered()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Method for initializing the object parameters during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$new(
  subset,
  heuristic,
  configuration = DependencyBasedStrategyConfiguration$new()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{The \link{Subset} used to apply the feature-clustering strategy.}

\item{\code{heuristic}}{The heuristic used to compute the relevance of each feature.
Must inherit from \link{GenericHeuristic} abstract class.}

\item{\code{configuration}}{optional paramter tocustomize configuration parameters for the strategy.
Must inherited from \link{StrategyConfiguration} abstract class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
Function responsible of performing the dependency-based feature clustering
strategy over the defined \link{Subset}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$execute(verbose = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{A logical value to specify if more verbosity is needed.}

\item{\code{...}}{Further arguments passed down to \code{execute} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDistribution"></a>}}
\subsection{Method \code{getDistribution()}}{
Function used to obtain a specific cluster distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$getDistribution(
  num.clusters = NULL,
  num.groups = NULL,
  include.unclustered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.clusters}}{A numeric value to select the number of clusters (define the distribution).}

\item{\code{num.groups}}{A single or \link{numeric} vector value to identify a specific group that
forms the clustering distribution.}

\item{\code{include.unclustered}}{A logical value to determine if unclustered features should be included.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{list} with the features comprising an specific clustering distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createTrain"></a>}}
\subsection{Method \code{createTrain()}}{
The function is used to create a \link{Trainset} object from a specific clustering distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$createTrain(
  subset,
  num.clusters = NULL,
  num.groups = NULL,
  include.unclustered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{The \link{Subset} object used as a basis to create the train set (see \link{Trainset} class).}

\item{\code{num.clusters}}{A \link{numeric} value to select the number of clusters (define the distribution).}

\item{\code{num.groups}}{A single or \link{numeric} vector value to identify a specific group that
forms the clustering distribution.}

\item{\code{include.unclustered}}{A logical value to determine if unclustered features should be included.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If num.clusters and num.groups are not defined, best clustering distribution is used to create the train set.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
The function is responsible for creating a plot to visualize the clustering distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$plot(dir.path = NULL, file.name = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{An optional argument to define the name of the directory where the exported plot will be saved.
If not defined, the file path will be automatically assigned to the current working directory, 'getwd()'.}

\item{\code{file.name}}{A character to define the name of the PDF file where the plot is exported.}

\item{\code{...}}{Further arguments passed down to \code{execute} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveCSV"></a>}}
\subsection{Method \code{saveCSV()}}{
The function is used to save the clustering distribution to a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$saveCSV(
  dir.path = NULL,
  name = NULL,
  num.clusters = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{The name of the directory to save the CSV file.}

\item{\code{name}}{Defines the name of the CSV file.}

\item{\code{num.clusters}}{An optional parameter to select the number of clusters to be saved.
If not defined, all cluster distributions will be saved.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
