% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/clustering.strategies.DependencyBasedStrategyConfiguration.R
\docType{class}
\name{DependencyBasedStrategyConfiguration}
\alias{DependencyBasedStrategyConfiguration}
\title{<<tittle>>}
\description{
DependencyBasedStrategyConfiguration
}
\details{
<<details>
}
\seealso{
\code{\link{StrategyConfiguration}}
}
\section{Super class}{
\code{\link[DDMCS:StrategyConfiguration]{DDMCS::StrategyConfiguration}} -> \code{DependencyBasedStrategyConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DependencyBasedStrategyConfiguration$new()}}
\item \href{#method-minNumClusters}{\code{DependencyBasedStrategyConfiguration$minNumClusters()}}
\item \href{#method-maxNumClusters}{\code{DependencyBasedStrategyConfiguration$maxNumClusters()}}
\item \href{#method-getBinaryCutoff}{\code{DependencyBasedStrategyConfiguration$getBinaryCutoff()}}
\item \href{#method-getRealCutoff}{\code{DependencyBasedStrategyConfiguration$getRealCutoff()}}
\item \href{#method-tiebreak}{\code{DependencyBasedStrategyConfiguration$tiebreak()}}
\item \href{#method-qualityOfCluster}{\code{DependencyBasedStrategyConfiguration$qualityOfCluster()}}
\item \href{#method-isImprovingClustering}{\code{DependencyBasedStrategyConfiguration$isImprovingClustering()}}
\item \href{#method-clone}{\code{DependencyBasedStrategyConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-minNumClusters"></a>}}
\subsection{Method \code{minNumClusters()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$minNumClusters(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-maxNumClusters"></a>}}
\subsection{Method \code{maxNumClusters()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$maxNumClusters(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBinaryCutoff"></a>}}
\subsection{Method \code{getBinaryCutoff()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$getBinaryCutoff()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRealCutoff"></a>}}
\subsection{Method \code{getRealCutoff()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$getRealCutoff()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tiebreak"></a>}}
\subsection{Method \code{tiebreak()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$tiebreak(
  feature,
  clus.candidates,
  fea.dep.dist.clus,
  corpus,
  heuristic,
  class,
  class.name
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{<<description>>}

\item{\code{clus.candidates}}{<<description>>}

\item{\code{fea.dep.dist.clus}}{<<description>>}

\item{\code{corpus}}{<<description>>}

\item{\code{heuristic}}{<<description>>}

\item{\code{class}}{<<description>>}

\item{\code{class.name}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-qualityOfCluster"></a>}}
\subsection{Method \code{qualityOfCluster()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$qualityOfCluster(clusters, metrics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusters}}{<<description>>}

\item{\code{metrics}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isImprovingClustering"></a>}}
\subsection{Method \code{isImprovingClustering()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$isImprovingClustering(clusters.deltha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusters.deltha}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
