% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.heuristics.GainRatioHeuristic.R
\docType{class}
\name{GainRatioHeuristic}
\alias{GainRatioHeuristic}
\title{Feature-clustering based on GainRatio methodology.}
\description{
Performs the feature-clustering using entropy-based filters.
}
\seealso{
\code{\link{Dataset}}, \code{\link[FSelector]{gain.ratio}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[DDMCS:GenericHeuristic]{DDMCS::GenericHeuristic}} -> \code{GainRatioHeuristic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GainRatioHeuristic$new()}}
\item \href{#method-heuristic}{\code{GainRatioHeuristic$heuristic()}}
\item \href{#method-clone}{\code{GainRatioHeuristic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates the \link{GainRatioHeuristic} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GainRatioHeuristic$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{GainRatioHeuristic} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-heuristic"></a>}}
\subsection{Method \code{heuristic()}}{
The algorithms find weights of discrete attributes basing on their
correlation with continous class attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GainRatioHeuristic$heuristic(col1, col2, column.names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{a numeric vector or matrix required to perform the
clustering operation.}

\item{\code{col2}}{a numeric vector or matrix to perform the clustering operation.}

\item{\code{column.names}}{an optional \link{character} vector with the names of
both columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \link{numeric} vector of length 1 or \link{NA} if an error occurs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GainRatioHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
