% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.heuristics.GenericHeuristic.R
\docType{class}
\name{GenericHeuristic}
\alias{GenericHeuristic}
\title{<<tittle>>}
\description{
GenericHeuristic
}
\details{
<<details>
}
\seealso{
\code{\link{Dataset}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenericHeuristic$new()}}
\item \href{#method-heuristic}{\code{GenericHeuristic$heuristic()}}
\item \href{#method-clone}{\code{GenericHeuristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericHeuristic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-heuristic"></a>}}
\subsection{Method \code{heuristic()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericHeuristic$heuristic(col1, col2, column.names = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{<<description>>}

\item{\code{col2}}{<<description>>}

\item{\code{column.names}}{<<description>>}

\item{\code{...}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
