% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.utility.GenericModelFit.R
\docType{class}
\name{GenericModelFit}
\alias{GenericModelFit}
\title{<<tittle>>}
\description{
GenericModelFit
}
\details{
<<details>
}
\seealso{
\code{\link{DefaultModelFit}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenericModelFit$new()}}
\item \href{#method-createFormula}{\code{GenericModelFit$createFormula()}}
\item \href{#method-createRecipe}{\code{GenericModelFit$createRecipe()}}
\item \href{#method-clone}{\code{GenericModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericModelFit$new(instances, class.name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instances}}{<<description>}

\item{\code{class.name}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createFormula"></a>}}
\subsection{Method \code{createFormula()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericModelFit$createFormula(simplify = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simplify}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createRecipe"></a>}}
\subsection{Method \code{createRecipe()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericModelFit$createRecipe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
