% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.plot.GenericPlot.R
\name{GenericPlot}
\alias{GenericPlot}
\title{Pseudo-abstract class for creating feature clustering plots.}
\description{
The \link{GenericPlot} implements a basic plot.
}
\seealso{
\code{\link{BinaryPlot}}
}
\keyword{color}
\keyword{device}
\keyword{hplot}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenericPlot$new()}}
\item \href{#method-plot}{\code{GenericPlot$plot()}}
\item \href{#method-clone}{\code{GenericPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
empty function used to initalize the object arguments in runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericPlot$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{GenericPlot} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
implements a generic plot to visualize basic feature-clustering data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericPlot$plot(summary, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary}}{a \link{data.frame} comprising the elements to be plotted.}

\item{\code{...}}{further arguments passed down to \code{plot} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
