% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.HDSubset.R
\docType{class}
\name{HDSubset}
\alias{HDSubset}
\title{<<tittle>>}
\description{
HDSubset
}
\details{
<<details>
}
\seealso{
\code{\link{Dataset}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HDSubset$new()}}
\item \href{#method-getFeatureNames}{\code{HDSubset$getFeatureNames()}}
\item \href{#method-getNcol}{\code{HDSubset$getNcol()}}
\item \href{#method-getID}{\code{HDSubset$getID()}}
\item \href{#method-getIterator}{\code{HDSubset$getIterator()}}
\item \href{#method-finalize}{\code{HDSubset$finalize()}}
\item \href{#method-isBlinded}{\code{HDSubset$isBlinded()}}
\item \href{#method-clone}{\code{HDSubset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDSubset$new(
  file.path,
  feature.names,
  feature.id,
  start.at = 0,
  sep = ",",
  chunk.size
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file.path}}{<<description>>}

\item{\code{feature.names}}{<<description>>}

\item{\code{feature.id}}{<<description>>}

\item{\code{start.at}}{<<description>>}

\item{\code{sep}}{<<description>>}

\item{\code{chunk.size}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeatureNames"></a>}}
\subsection{Method \code{getFeatureNames()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDSubset$getFeatureNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNcol"></a>}}
\subsection{Method \code{getNcol()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDSubset$getNcol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\subsection{Method \code{getID()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDSubset$getID()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIterator"></a>}}
\subsection{Method \code{getIterator()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDSubset$getIterator(chunk.size = private$chunk.size, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chunk.size}}{<<description>>}

\item{\code{verbose}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDSubset$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isBlinded"></a>}}
\subsection{Method \code{isBlinded()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDSubset$isBlinded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDSubset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
