% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.MaximizeFN.R
\docType{class}
\name{MaximizeFN}
\alias{MaximizeFN}
\title{Combined metric strategy to maximize FN errors.}
\description{
Calculates if the positive class is the predicted one in any of
the metrics, otherwise, the instance is not considered to have the positive
class associated.
}
\seealso{
\code{\link{CombinedMetrics}}
}
\keyword{classif}
\keyword{math}
\section{Super class}{
\code{\link[DDMCS:CombinedMetrics]{DDMCS::CombinedMetrics}} -> \code{MaximizeFN}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MaximizeFN$new()}}
\item \href{#method-getFinalPrediction}{\code{MaximizeFN$getFinalPrediction()}}
\item \href{#method-clone}{\code{MaximizeFN$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="CombinedMetrics" data-id="getRequiredMetrics">}\href{../../DDMCS/html/CombinedMetrics.html#method-getRequiredMetrics}{\code{DDMCS::CombinedMetrics$getRequiredMetrics()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaximizeFN$new(required.metrics = c("MCC", "PPV"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{required.metrics}}{A \code{\link{character}} vector of length 1
with the name of the required metrics.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFinalPrediction"></a>}}
\subsection{Method \code{getFinalPrediction()}}{
Function to obtain the final prediction based on different
metrics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaximizeFN$getFinalPrediction(
  raw.pred,
  prob.pred,
  positive.class,
  negative.class
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw.pred}}{A \code{\link{character}} list of length greater than 2
with the class value of the predictions made by the metrics.}

\item{\code{prob.pred}}{A \code{\link{numeric}} list of length greater than 2
with the probability of the predictions made by the metrics.}

\item{\code{positive.class}}{A \code{\link{character}} with the value of the
positive class.}

\item{\code{negative.class}}{A \code{\link{character}} with the value of the
negative class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{logical} value indicating if the instance is predicted
as positive class or not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaximizeFN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
