% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.Methodology.R
\docType{class}
\name{Methodology}
\alias{Methodology}
\title{<<miguel>>}
\description{
Methodology
}
\details{
<<details>
}
\seealso{
\code{\link{ProbBasedMethodology}}
}
\keyword{math}
\keyword{misc}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Methodology$new()}}
\item \href{#method-getRequiredMetrics}{\code{Methodology$getRequiredMetrics()}}
\item \href{#method-compute}{\code{Methodology$compute()}}
\item \href{#method-clone}{\code{Methodology$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methodology$new(required.metrics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{required.metrics}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRequiredMetrics"></a>}}
\subsection{Method \code{getRequiredMetrics()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methodology$getRequiredMetrics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\subsection{Method \code{compute()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methodology$compute(raw.pred, prob.pred, positive.class, negative.class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw.pred}}{<<description>>}

\item{\code{prob.pred}}{<<description>>}

\item{\code{positive.class}}{<<description>>}

\item{\code{negative.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methodology$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
