% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.MinimizeFP.R
\docType{class}
\name{MinimizeFP}
\alias{MinimizeFP}
\title{<<tittle>>}
\description{
MinimizeFP
}
\details{
<<details>
}
\seealso{
\code{\link{CombinedMetrics}}
}
\section{Super class}{
\code{\link[DDMCS:CombinedMetrics]{DDMCS::CombinedMetrics}} -> \code{MinimizeFP}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MinimizeFP$new()}}
\item \href{#method-getFinalPrediction}{\code{MinimizeFP$getFinalPrediction()}}
\item \href{#method-clone}{\code{MinimizeFP$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="CombinedMetrics" data-id="getRequiredMetrics">}\href{../../DDMCS/html/CombinedMetrics.html#method-getRequiredMetrics}{\code{DDMCS::CombinedMetrics$getRequiredMetrics()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MinimizeFP$new(required.metrics = c("MCC", "PPV"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{required.metrics}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFinalPrediction"></a>}}
\subsection{Method \code{getFinalPrediction()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MinimizeFP$getFinalPrediction(
  raw.pred,
  prob.pred,
  positive.class,
  negative.class
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw.pred}}{<<description>>}

\item{\code{prob.pred}}{<<description>>}

\item{\code{positive.class}}{<<description>>}

\item{\code{negative.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MinimizeFP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
