% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.summaryFunction.NoProbability.R
\docType{class}
\name{NoProbability}
\alias{NoProbability}
\title{<<tittle>>}
\description{
NoProbability
}
\details{
<<details>
}
\seealso{
\code{\link{SummaryFunction}}
}
\section{Super class}{
\code{\link[DDMCS:SummaryFunction]{DDMCS::SummaryFunction}} -> \code{NoProbability}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NoProbability$new()}}
\item \href{#method-execute}{\code{NoProbability$execute()}}
\item \href{#method-clone}{\code{NoProbability$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SummaryFunction" data-id="getMeasures">}\href{../../DDMCS/html/SummaryFunction.html#method-getMeasures}{\code{DDMCS::SummaryFunction$getMeasures()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{

\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$execute(data, lev = NULL, model = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{<<description>>}

\item{\code{lev}}{<<description>>}

\item{\code{model}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
