% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.heuristics.OddsRatioHeuristic.R
\docType{class}
\name{OddsRatioHeuristic}
\alias{OddsRatioHeuristic}
\title{<<tittle>>}
\description{
OddsRatioHeuristic
}
\details{
<<details>
}
\seealso{
\code{\link{Dataset}}
}
\section{Super class}{
\code{\link[DDMCS:GenericHeuristic]{DDMCS::GenericHeuristic}} -> \code{OddsRatioHeuristic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OddsRatioHeuristic$new()}}
\item \href{#method-heuristic}{\code{OddsRatioHeuristic$heuristic()}}
\item \href{#method-clone}{\code{OddsRatioHeuristic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OddsRatioHeuristic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-heuristic"></a>}}
\subsection{Method \code{heuristic()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OddsRatioHeuristic$heuristic(col1, col2, column.names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{<<description>>}

\item{\code{col2}}{<<description>>}

\item{\code{column.names}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OddsRatioHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
