% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.ProbAverageWeightedVoting.R
\name{ProbAverageWeightedVoting}
\alias{ProbAverageWeightedVoting}
\title{Implementation of Probabilistic Average Weighted voting.}
\description{
Computes the final prediction by perfoming the weighted mean of
the probability achieved by each cluster prediction. By default,
weight values are consistent with the performance value
achieved by the best M.L. model on each cluster.
}
\seealso{
\code{\link{DDMCS}}, \code{\link{ClassMajorityVoting}}, \code{\link{ClassWeightedVoting}},
\code{\link{ProbAverageVoting}}, \code{\link{ProbAverageWeightedVoting}}, \code{\link{ProbBasedMethodology}}
}
\keyword{math}
\keyword{methods}
\keyword{models}
\section{Super class}{
\code{\link[DDMCS:SimpleVoting]{DDMCS::SimpleVoting}} -> \code{ProbAverageWeightedVoting}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProbAverageWeightedVoting$new()}}
\item \href{#method-getWeights}{\code{ProbAverageWeightedVoting$getWeights()}}
\item \href{#method-setWeights}{\code{ProbAverageWeightedVoting$setWeights()}}
\item \href{#method-execute}{\code{ProbAverageWeightedVoting$execute()}}
\item \href{#method-clone}{\code{ProbAverageWeightedVoting$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SimpleVoting" data-id="getCutoff">}\href{../../DDMCS/html/SimpleVoting.html#method-getCutoff}{\code{DDMCS::SimpleVoting$getCutoff()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SimpleVoting" data-id="getFinalPred">}\href{../../DDMCS/html/SimpleVoting.html#method-getFinalPred}{\code{DDMCS::SimpleVoting$getFinalPred()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageWeightedVoting$new(cutoff = 0.5, weights = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cutoff}}{A \code{\link{character}} vector defining the minimum probability used to perform a
a positive classification. If is not defined, 0.5 will be used as default value.}

\item{\code{weights}}{A \code{\link{numeric}} vector with the weigths of each cluster. If \code{\link{NULL}} performance
achieved during training will be used as default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWeights"></a>}}
\subsection{Method \code{getWeights()}}{
The function returns the value of the majority class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageWeightedVoting$getWeights()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link{character}} vector of length 1 with the name of the majority class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setWeights"></a>}}
\subsection{Method \code{setWeights()}}{
The funtion allows changing the value of the weights.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageWeightedVoting$setWeights(weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weights}}{A \code{\link{numeric}} vector containing the new weights.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
The function implements the cluster-weighted probabilistic voting procedure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageWeightedVoting$execute(predictions, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{A \code{\link{ClusterPredictions}} object containing all
the predictions achieved for each cluster.}

\item{\code{verbose}}{A \code{\link{logical}} value to specify if more verbosity is needed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageWeightedVoting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
