% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.Sensitivity.R
\name{Sensitivity}
\alias{Sensitivity}
\title{Computes the Sensitivity Value.}
\description{
Sensitivity is a measure of the proportion of actual positive cases that got predicted as positive (or true positive).
}
\details{
\deqn{Sensitivity = \frac{(TP)}{(TP + FN)}}
}
\seealso{
\code{\link{MeasureFunction}}, \code{\link{ClassificationOutput}}, \code{\link{ConfMatrix}}

\code{\link{ConfMatrix}}
}
\keyword{classif}
\keyword{math}
\section{Super class}{
\code{\link[DDMCS:MeasureFunction]{DDMCS::MeasureFunction}} -> \code{Sensitivity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sensitivity$new()}}
\item \href{#method-compute}{\code{Sensitivity$compute()}}
\item \href{#method-clone}{\code{Sensitivity$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sensitivity$new(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter to
define the type of object used as basis to compute the \code{Sensitivity} measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{\link{Sensitivity}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\subsection{Method \code{compute()}}{
The function computes the \code{Sensitivity} achieved by the M.L. model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sensitivity$compute(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter to define the
type of object used as basis to compute the \code{Sensitivity} measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is automatically invoque by the \link{ClassificationOutput} object.
}

\subsection{Returns}{
A \code{\link{numeric}} vector of size 1 or \code{\link{NULL}} if an error occured.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sensitivity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
