% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.Specificity.R
\docType{class}
\name{Specificity}
\alias{Specificity}
\title{<<tittle>>}
\description{
Specificity
}
\details{
<<details>
}
\seealso{
\code{\link{MeasureFunction}}
}
\section{Super class}{
\code{\link[DDMCS:MeasureFunction]{DDMCS::MeasureFunction}} -> \code{Specificity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Specificity$new()}}
\item \href{#method-compute}{\code{Specificity$compute()}}
\item \href{#method-clone}{\code{Specificity$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specificity$new(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\subsection{Method \code{compute()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specificity$compute(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specificity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
