% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.trainFunctions.TrainFunction.R
\docType{class}
\name{TrainFunction}
\alias{TrainFunction}
\title{<<tittle>>}
\description{
TrainFunction
}
\details{
<<details>
}
\seealso{
\code{\link{TwoClass}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TrainFunction$new()}}
\item \href{#method-create}{\code{TrainFunction$create()}}
\item \href{#method-getResamplingMethod}{\code{TrainFunction$getResamplingMethod()}}
\item \href{#method-getNumberFolds}{\code{TrainFunction$getNumberFolds()}}
\item \href{#method-getSavePredictions}{\code{TrainFunction$getSavePredictions()}}
\item \href{#method-getClassProbs}{\code{TrainFunction$getClassProbs()}}
\item \href{#method-getAllowParallel}{\code{TrainFunction$getAllowParallel()}}
\item \href{#method-getVerboseIter}{\code{TrainFunction$getVerboseIter()}}
\item \href{#method-getTrFunction}{\code{TrainFunction$getTrFunction()}}
\item \href{#method-getMeasures}{\code{TrainFunction$getMeasures()}}
\item \href{#method-getType}{\code{TrainFunction$getType()}}
\item \href{#method-getSeed}{\code{TrainFunction$getSeed()}}
\item \href{#method-setSummaryFunction}{\code{TrainFunction$setSummaryFunction()}}
\item \href{#method-setClassProbs}{\code{TrainFunction$setClassProbs()}}
\item \href{#method-clone}{\code{TrainFunction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$new(
  method,
  number,
  savePredictions,
  classProbs,
  allowParallel,
  verboseIter,
  seed
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{<<description>>}

\item{\code{number}}{<<description>>}

\item{\code{savePredictions}}{<<description>>}

\item{\code{classProbs}}{<<description>>}

\item{\code{allowParallel}}{<<description>>}

\item{\code{verboseIter}}{<<description>>}

\item{\code{seed}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\subsection{Method \code{create()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$create(summaryFunction, search.method = "grid", class.probs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summaryFunction}}{<<description>>}

\item{\code{search.method}}{<<description>>}

\item{\code{class.probs}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResamplingMethod"></a>}}
\subsection{Method \code{getResamplingMethod()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getResamplingMethod()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNumberFolds"></a>}}
\subsection{Method \code{getNumberFolds()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getNumberFolds()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSavePredictions"></a>}}
\subsection{Method \code{getSavePredictions()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getSavePredictions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassProbs"></a>}}
\subsection{Method \code{getClassProbs()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getClassProbs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAllowParallel"></a>}}
\subsection{Method \code{getAllowParallel()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getAllowParallel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVerboseIter"></a>}}
\subsection{Method \code{getVerboseIter()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getVerboseIter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTrFunction"></a>}}
\subsection{Method \code{getTrFunction()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getTrFunction()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMeasures"></a>}}
\subsection{Method \code{getMeasures()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getMeasures()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getType"></a>}}
\subsection{Method \code{getType()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSeed"></a>}}
\subsection{Method \code{getSeed()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$getSeed()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSummaryFunction"></a>}}
\subsection{Method \code{setSummaryFunction()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$setSummaryFunction(summaryFunction)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summaryFunction}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setClassProbs"></a>}}
\subsection{Method \code{setClassProbs()}}{
<<description>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$setClassProbs(class.probs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class.probs}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainFunction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
