% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.TrainOutput.R
\docType{class}
\name{TrainOutput}
\alias{TrainOutput}
\title{<<tittle>>}
\description{
TrainOutput
}
\details{
<<details>
}
\seealso{
\code{\link{DDMCS}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TrainOutput$new()}}
\item \href{#method-getModels}{\code{TrainOutput$getModels()}}
\item \href{#method-getPerformance}{\code{TrainOutput$getPerformance()}}
\item \href{#method-savePerformance}{\code{TrainOutput$savePerformance()}}
\item \href{#method-plot}{\code{TrainOutput$plot()}}
\item \href{#method-getMetrics}{\code{TrainOutput$getMetrics()}}
\item \href{#method-getClassValues}{\code{TrainOutput$getClassValues()}}
\item \href{#method-getPositiveClass}{\code{TrainOutput$getPositiveClass()}}
\item \href{#method-getSize}{\code{TrainOutput$getSize()}}
\item \href{#method-clone}{\code{TrainOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$new(models, class.values, positive.class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{models}}{<<descrription>>}

\item{\code{class.values}}{<<descrription>>}

\item{\code{positive.class}}{<<descrription>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getModels"></a>}}
\subsection{Method \code{getModels()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$getModels(metric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric}}{<<descrription>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPerformance"></a>}}
\subsection{Method \code{getPerformance()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$getPerformance(metrics = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metrics}}{<<descrription>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-savePerformance"></a>}}
\subsection{Method \code{savePerformance()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$savePerformance(dir.path, metrics = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{<<description>>}

\item{\code{metrics}}{<<descrription>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$plot(dir.path, metrics = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{<<description>>}

\item{\code{metrics}}{<<descrription>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetrics"></a>}}
\subsection{Method \code{getMetrics()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$getMetrics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassValues"></a>}}
\subsection{Method \code{getClassValues()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$getClassValues()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPositiveClass"></a>}}
\subsection{Method \code{getPositiveClass()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$getPositiveClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSize"></a>}}
\subsection{Method \code{getSize()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$getSize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
