% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.strategies.BinaryRealTypeStrategy.R
\docType{class}
\name{BinaryRealTypeStrategy}
\alias{BinaryRealTypeStrategy}
\title{<<tittle>>}
\description{
BinaryRealTypeStrategy
}
\details{
<<details>
}
\seealso{
\code{\link{ClusteringStrategy}}
}
\section{Super class}{
\code{\link[DDMCS:ClusteringStrategy]{DDMCS::ClusteringStrategy}} -> \code{BinaryRealTypeStrategy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BinaryRealTypeStrategy$new()}}
\item \href{#method-execute}{\code{BinaryRealTypeStrategy$execute()}}
\item \href{#method-getDistribution}{\code{BinaryRealTypeStrategy$getDistribution()}}
\item \href{#method-createTrain}{\code{BinaryRealTypeStrategy$createTrain()}}
\item \href{#method-plot}{\code{BinaryRealTypeStrategy$plot()}}
\item \href{#method-saveCSV}{\code{BinaryRealTypeStrategy$saveCSV()}}
\item \href{#method-clone}{\code{BinaryRealTypeStrategy$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getBestClusterDistribution">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getBestClusterDistribution}{\code{DDMCS::ClusteringStrategy$getBestClusterDistribution()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getConfiguration">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getConfiguration}{\code{DDMCS::ClusteringStrategy$getConfiguration()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getDescription">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getDescription}{\code{DDMCS::ClusteringStrategy$getDescription()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getHeuristic">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getHeuristic}{\code{DDMCS::ClusteringStrategy$getHeuristic()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="ClusteringStrategy" data-id="getUnclustered">}\href{../../DDMCS/html/ClusteringStrategy.html#method-getUnclustered}{\code{DDMCS::ClusteringStrategy$getUnclustered()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRealTypeStrategy$new(
  subset,
  heuristic,
  configuration = StrategyConfiguration$new()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{<<description>>}

\item{\code{heuristic}}{<<description>>}

\item{\code{configuration}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRealTypeStrategy$execute(verbose = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{<<description>>}

\item{\code{...}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDistribution"></a>}}
\subsection{Method \code{getDistribution()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRealTypeStrategy$getDistribution(
  num.clusters = NULL,
  num.groups = NULL,
  include.unclustered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.clusters}}{<<description>>}

\item{\code{num.groups}}{<<description>>}

\item{\code{include.unclustered}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createTrain"></a>}}
\subsection{Method \code{createTrain()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRealTypeStrategy$createTrain(
  subset,
  num.clusters = NULL,
  num.groups = NULL,
  include.unclustered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{<<description>>}

\item{\code{num.clusters}}{<<description>>}

\item{\code{num.groups}}{<<description>>}

\item{\code{include.unclustered}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRealTypeStrategy$plot(dir.path = NULL, file.name = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{<<description>>}

\item{\code{file.name}}{<<description>>}

\item{\code{...}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveCSV"></a>}}
\subsection{Method \code{saveCSV()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRealTypeStrategy$saveCSV(
  dir.path = NULL,
  name = NULL,
  num.clusters = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{<<description>>}

\item{\code{name}}{<<description>>}

\item{\code{num.clusters}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRealTypeStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
