% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.heuristics.ChiSquareHeuristic.R
\name{ChiSquareHeuristic}
\alias{ChiSquareHeuristic}
\title{Feature-clustering based on ChiSquare method.}
\description{
Performs feature-clustering based on ChiSquare method.
}
\seealso{
\code{\link{Dataset}}, \code{\link[stats]{chisq.test}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[D2MCS:GenericHeuristic]{D2MCS::GenericHeuristic}} -> \code{ChiSquareHeuristic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChiSquareHeuristic-new}{\code{ChiSquareHeuristic$new()}}
\item \href{#method-ChiSquareHeuristic-heuristic}{\code{ChiSquareHeuristic$heuristic()}}
\item \href{#method-ChiSquareHeuristic-clone}{\code{ChiSquareHeuristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChiSquareHeuristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChiSquareHeuristic-new}{}}}
\subsection{Method \code{new()}}{
Empty function used to initialize the object arguments in
runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChiSquareHeuristic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChiSquareHeuristic-heuristic"></a>}}
\if{latex}{\out{\hypertarget{method-ChiSquareHeuristic-heuristic}{}}}
\subsection{Method \code{heuristic()}}{
Functions responsible of performing the ChiSquare
feature-clustering operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChiSquareHeuristic$heuristic(col1, col2, column.names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{A \link{numeric} vector or matrix required to perform the
clustering operation.}

\item{\code{col2}}{A \link{numeric} vector or matrix to perform the clustering
operation.}

\item{\code{column.names}}{An optional \link{character} vector with the names of
both columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1 or \link{NA} if an error
occurs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChiSquareHeuristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChiSquareHeuristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChiSquareHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
