% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.strategies.ClusteringStrategy.R
\docType{class}
\name{ClusteringStrategy}
\alias{ClusteringStrategy}
\title{Abstract Feature Clustering Strategy class.}
\description{
Abstract class used as a template to ensure the proper definition
of new customized clustering strategies.
}
\details{
The \link{ClusteringStrategy} is an archetype class so it cannot be instantiated.
}
\seealso{
\code{\link{Subset}}, \code{\link{GenericHeuristic}}
}
\keyword{cluster}
\keyword{manip}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClusteringStrategy$new()}}
\item \href{#method-getDescription}{\code{ClusteringStrategy$getDescription()}}
\item \href{#method-getHeuristic}{\code{ClusteringStrategy$getHeuristic()}}
\item \href{#method-getConfiguration}{\code{ClusteringStrategy$getConfiguration()}}
\item \href{#method-getBestClusterDistribution}{\code{ClusteringStrategy$getBestClusterDistribution()}}
\item \href{#method-getUnclustered}{\code{ClusteringStrategy$getUnclustered()}}
\item \href{#method-execute}{\code{ClusteringStrategy$execute()}}
\item \href{#method-getDistribution}{\code{ClusteringStrategy$getDistribution()}}
\item \href{#method-createTrain}{\code{ClusteringStrategy$createTrain()}}
\item \href{#method-plot}{\code{ClusteringStrategy$plot()}}
\item \href{#method-saveCSV}{\code{ClusteringStrategy$saveCSV()}}
\item \href{#method-clone}{\code{ClusteringStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
A function responsible for creating a \link{ClusteringStrategy} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$new(subset, heuristic, description, configuration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{a \link{Subset} object to perform the clustering strategy.}

\item{\code{heuristic}}{the heuristic to be applied. Must inherite from \link{GenericHeuristic} class.}

\item{\code{description}}{a \link{character} vector describing the strategy operation.}

\item{\code{configuration}}{optional customized configuration parameters for the strategy.
Must inherited from \link{StrategyConfiguration} abstract class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDescription"></a>}}
\subsection{Method \code{getDescription()}}{
the function is used to obtain the description of the strategie.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$getDescription()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{character} vector of \link{NULL} if not defined.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHeuristic"></a>}}
\subsection{Method \code{getHeuristic()}}{
the function returns the heuristic applied for the clustering strategy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$getHeuristic()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object inherited from \link{ClusteringStrategy} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConfiguration"></a>}}
\subsection{Method \code{getConfiguration()}}{
the function returns the configuration parameters used to perform the clustering strategy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$getConfiguration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object inherited from \link{StrategyConfiguration} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBestClusterDistribution"></a>}}
\subsection{Method \code{getBestClusterDistribution()}}{
the function obtains the best clustering distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$getBestClusterDistribution()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{list} of clusters. Each list element represents a feature group.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUnclustered"></a>}}
\subsection{Method \code{getUnclustered()}}{
the function is used to return the features that cannot be clustered due to
incompatibilities with the used heuristic.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$getUnclustered()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{character} vector containing the unclassified features.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
abstract function responsible of performing the clustering
strategy over the defined \link{Subset}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$execute(verbose, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{a logical value to specify if more verbosity is needed.}

\item{\code{...}}{further arguments passed down to \code{execute} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDistribution"></a>}}
\subsection{Method \code{getDistribution()}}{
abstract function used to obtain the set of features following
an specific clustering distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$getDistribution(
  num.clusters = NULL,
  num.groups = NULL,
  include.unclustered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.clusters}}{a numeric value to select the number of clusters (define the distribution).}

\item{\code{num.groups}}{a single or \link{numeric} vector value to identify a specific group that
forms the clustering distribution.}

\item{\code{include.unclustered}}{a logical value to determine if unclustered features should be included.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \link{list} with the features comprising an specific clustering distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createTrain"></a>}}
\subsection{Method \code{createTrain()}}{
Abstract function in charge of creating a \link{Trainset} object for training purposes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$createTrain(
  subset,
  num.cluster = NULL,
  num.groups = NULL,
  include.unclustered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{A \link{Subset} object used as a basis to create the \link{Trainset}}

\item{\code{num.cluster}}{A \link{numeric} value to select the number of clusters (define the distribution).}

\item{\code{num.groups}}{A single or \link{numeric} vector value to identify a specific group that
forms the clustering distribution.}

\item{\code{include.unclustered}}{A logical value to determine if unclustered features should be included.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Abstract function responsible of creating a plot to visualize the clustering distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$plot(dir.path = NULL, file.name = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{An optional argument to define the name of the directory where the exported plot will be saved.
If not defined, the file path will be automatically assigned to the current working directory, ‘getwd()’.}

\item{\code{file.name}}{The name of the PDF file where the plot is exported.}

\item{\code{...}}{Further arguments passed down to \code{execute} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveCSV"></a>}}
\subsection{Method \code{saveCSV()}}{
abstract function to save the clustering distribution to a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$saveCSV(dir.path, name, num.clusters = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{The name of the directory to save the CSV file.}

\item{\code{name}}{Defines the name of the CSV file.}

\item{\code{num.clusters}}{An optional parameter to select the number of clusters to be saved.
If not defined, all clusters will be saved.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
