% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.Dataset.R
\docType{class}
\name{Dataset}
\alias{Dataset}
\title{<<tittle>>}
\description{
Dataset
}
\details{
<<details>
}
\seealso{
\code{\link{HDDataset}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dataset$new()}}
\item \href{#method-getColumnNames}{\code{Dataset$getColumnNames()}}
\item \href{#method-getDataset}{\code{Dataset$getDataset()}}
\item \href{#method-getClassName}{\code{Dataset$getClassName()}}
\item \href{#method-getClassIndex}{\code{Dataset$getClassIndex()}}
\item \href{#method-getClassValues}{\code{Dataset$getClassValues()}}
\item \href{#method-getPositiveClass}{\code{Dataset$getPositiveClass()}}
\item \href{#method-getNcol}{\code{Dataset$getNcol()}}
\item \href{#method-getNrow}{\code{Dataset$getNrow()}}
\item \href{#method-getClassSummary}{\code{Dataset$getClassSummary()}}
\item \href{#method-getRemovedColumns}{\code{Dataset$getRemovedColumns()}}
\item \href{#method-setPositiveClass}{\code{Dataset$setPositiveClass()}}
\item \href{#method-setClassIndex}{\code{Dataset$setClassIndex()}}
\item \href{#method-setClassName}{\code{Dataset$setClassName()}}
\item \href{#method-cleanData}{\code{Dataset$cleanData()}}
\item \href{#method-removeColumns}{\code{Dataset$removeColumns()}}
\item \href{#method-createPartitions}{\code{Dataset$createPartitions()}}
\item \href{#method-createSubset}{\code{Dataset$createSubset()}}
\item \href{#method-createTrain}{\code{Dataset$createTrain()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(
  filepath,
  header = TRUE,
  sep = ",",
  skip = 0,
  target.class,
  positive.class,
  normalize.names = FALSE,
  string.as.factor = FALSE,
  ignore.columns = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filepath}}{<<description>>}

\item{\code{header}}{<<description>>}

\item{\code{sep}}{<<description>>}

\item{\code{skip}}{<<description>>}

\item{\code{target.class}}{<<description>>}

\item{\code{positive.class}}{<<description>>}

\item{\code{normalize.names}}{<<description>>}

\item{\code{string.as.factor}}{<<description>>}

\item{\code{ignore.columns}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getColumnNames"></a>}}
\subsection{Method \code{getColumnNames()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getColumnNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDataset"></a>}}
\subsection{Method \code{getDataset()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getDataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassName"></a>}}
\subsection{Method \code{getClassName()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getClassName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassIndex"></a>}}
\subsection{Method \code{getClassIndex()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getClassIndex()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassValues"></a>}}
\subsection{Method \code{getClassValues()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getClassValues()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPositiveClass"></a>}}
\subsection{Method \code{getPositiveClass()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getPositiveClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNcol"></a>}}
\subsection{Method \code{getNcol()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getNcol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNrow"></a>}}
\subsection{Method \code{getNrow()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getNrow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassSummary"></a>}}
\subsection{Method \code{getClassSummary()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getClassSummary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRemovedColumns"></a>}}
\subsection{Method \code{getRemovedColumns()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$getRemovedColumns()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPositiveClass"></a>}}
\subsection{Method \code{setPositiveClass()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$setPositiveClass(positive.class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{positive.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setClassIndex"></a>}}
\subsection{Method \code{setClassIndex()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$setClassIndex(class.index, positive.class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class.index}}{<<description>>}

\item{\code{positive.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setClassName"></a>}}
\subsection{Method \code{setClassName()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$setClassName(class.name, positive.class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class.name}}{<<description>>}

\item{\code{positive.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\subsection{Method \code{cleanData()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$cleanData(remove.funcs = NULL, remove.na = TRUE, remove.const = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{remove.funcs}}{<<description>>}

\item{\code{remove.na}}{<<description>>}

\item{\code{remove.const}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeColumns"></a>}}
\subsection{Method \code{removeColumns()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$removeColumns(
  columns,
  remove.funcs = NULL,
  remove.na = FALSE,
  remove.const = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{<<description>>}

\item{\code{remove.funcs}}{<<description>>}

\item{\code{remove.na}}{<<description>>}

\item{\code{remove.const}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createPartitions"></a>}}
\subsection{Method \code{createPartitions()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$createPartitions(
  num.folds = NULL,
  percent.folds = NULL,
  class.balance = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.folds}}{<<description>>}

\item{\code{percent.folds}}{<<description>>}

\item{\code{class.balance}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createSubset"></a>}}
\subsection{Method \code{createSubset()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$createSubset(
  num.folds,
  column.id = NULL,
  opts = list(remove.na = TRUE, remove.const = FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.folds}}{<<description>>}

\item{\code{column.id}}{<<description>>}

\item{\code{opts}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createTrain"></a>}}
\subsection{Method \code{createTrain()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$createTrain(
  num.folds = NULL,
  opts = list(remove.na = TRUE, remove.const = FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.folds}}{<<description>>}

\item{\code{opts}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<return>>
}
}
}
