% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.strategies.DependencyBasedStrategy.R
\name{DependencyBasedStrategy}
\alias{DependencyBasedStrategy}
\title{Clustering strategy based on dependency between features.}
\description{
Features are distributed according to their independence values.
This strategy is divided into two steps. The first phase focuses on forming
groups with those features most dependent on each other. This step also
identifies those that are independent from all the others in the group.
The second step is to try out different numbers of clusters until you find
the one you think is best. These clusters are formed by inserting in all the
independent characteristics identified previously and trying to distribute
the features of the groups formed in the previous step in separate clusters.
In this way, it seeks to ensure that the features are as independent as
possible from those found in the same cluster.
}
\details{
The strategy is suitable only for binary and real features. Other
features are automatically grouped into a specific cluster named as
'unclustered'. This class requires the \code{\link{StrategyConfiguration}}
type object implements the following methods:

- \code{getBinaryCutoff()}: The function is used to define the interval to
consider the dependency between binary features.

- \code{getRealCutoff()}: The function allows defining the cutoff to consider
the dependency between real features.

- \code{tiebreak(feature, clus.candidates, fea.dep.dist.clus, corpus,
heuristic, class, class.name)}: The function solves the ties between two
(or more) features.

- \code{qualityOfCluster(clusters, metrics)}: The function determines the
quality of a cluster

- \code{isImprovingClustering(clusters.deltha)}: The function indicates if
clustering is getting better as the number of them increases.

An example of implementation with the description of each parameter is the
\code{\link{DependencyBasedStrategyConfiguration}} class.
}
\seealso{
\code{\link{GenericClusteringStrategy}},
\code{\link{StrategyConfiguration}},
\code{\link{DependencyBasedStrategyConfiguration}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[D2MCS:GenericClusteringStrategy]{D2MCS::GenericClusteringStrategy}} -> \code{DependencyBasedStrategy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DependencyBasedStrategy-new}{\code{DependencyBasedStrategy$new()}}
\item \href{#method-DependencyBasedStrategy-execute}{\code{DependencyBasedStrategy$execute()}}
\item \href{#method-DependencyBasedStrategy-getDistribution}{\code{DependencyBasedStrategy$getDistribution()}}
\item \href{#method-DependencyBasedStrategy-createTrain}{\code{DependencyBasedStrategy$createTrain()}}
\item \href{#method-DependencyBasedStrategy-plot}{\code{DependencyBasedStrategy$plot()}}
\item \href{#method-DependencyBasedStrategy-saveCSV}{\code{DependencyBasedStrategy$saveCSV()}}
\item \href{#method-DependencyBasedStrategy-clone}{\code{DependencyBasedStrategy$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="D2MCS" data-topic="GenericClusteringStrategy" data-id="getBestClusterDistribution"><a href='../../D2MCS/html/GenericClusteringStrategy.html#method-GenericClusteringStrategy-getBestClusterDistribution'><code>D2MCS::GenericClusteringStrategy$getBestClusterDistribution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="D2MCS" data-topic="GenericClusteringStrategy" data-id="getConfiguration"><a href='../../D2MCS/html/GenericClusteringStrategy.html#method-GenericClusteringStrategy-getConfiguration'><code>D2MCS::GenericClusteringStrategy$getConfiguration()</code></a></span></li>
<li><span class="pkg-link" data-pkg="D2MCS" data-topic="GenericClusteringStrategy" data-id="getDescription"><a href='../../D2MCS/html/GenericClusteringStrategy.html#method-GenericClusteringStrategy-getDescription'><code>D2MCS::GenericClusteringStrategy$getDescription()</code></a></span></li>
<li><span class="pkg-link" data-pkg="D2MCS" data-topic="GenericClusteringStrategy" data-id="getHeuristic"><a href='../../D2MCS/html/GenericClusteringStrategy.html#method-GenericClusteringStrategy-getHeuristic'><code>D2MCS::GenericClusteringStrategy$getHeuristic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="D2MCS" data-topic="GenericClusteringStrategy" data-id="getUnclustered"><a href='../../D2MCS/html/GenericClusteringStrategy.html#method-GenericClusteringStrategy-getUnclustered'><code>D2MCS::GenericClusteringStrategy$getUnclustered()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyBasedStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyBasedStrategy-new}{}}}
\subsection{Method \code{new()}}{
Method for initializing the object parameters during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$new(
  subset,
  heuristic,
  configuration = DependencyBasedStrategyConfiguration$new()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{The \code{\link{Subset}} used to apply the feature-clustering
strategy.}

\item{\code{heuristic}}{The heuristic used to compute the relevance of each
feature. Must inherit from \code{\link{GenericHeuristic}} abstract class.}

\item{\code{configuration}}{optional parameter to customize configuration
parameters for the strategy. Must inherited from
\code{\link{StrategyConfiguration}} abstract class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyBasedStrategy-execute"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyBasedStrategy-execute}{}}}
\subsection{Method \code{execute()}}{
Function responsible of performing the dependency-based
feature clustering strategy over the defined \code{\link{Subset}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$execute(verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{A \link{logical} value to specify if more verbosity is
needed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyBasedStrategy-getDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyBasedStrategy-getDistribution}{}}}
\subsection{Method \code{getDistribution()}}{
Function used to obtain a specific cluster distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$getDistribution(
  num.clusters = NULL,
  num.groups = NULL,
  include.unclustered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.clusters}}{A \link{numeric} value to select the number of
clusters (define the distribution).}

\item{\code{num.groups}}{A single or \link{numeric} vector value to identify a
specific group that forms the clustering distribution.}

\item{\code{include.unclustered}}{A \link{logical} value to determine if
unclustered features should be included.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{list} with the features comprising an specific
clustering distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyBasedStrategy-createTrain"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyBasedStrategy-createTrain}{}}}
\subsection{Method \code{createTrain()}}{
The function is used to create a \code{\link{Trainset}}
object from a specific clustering distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$createTrain(
  subset,
  num.clusters = NULL,
  num.groups = NULL,
  include.unclustered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{The \code{\link{Subset}} object used as a basis to create
the train set (see \code{\link{Trainset}} class).}

\item{\code{num.clusters}}{A \link{numeric} value to select the number of
clusters (define the distribution).}

\item{\code{num.groups}}{A single or \link{numeric} vector value to identify a
specific group that forms the clustering distribution.}

\item{\code{include.unclustered}}{A \link{logical} value to determine if
unclustered features should be included.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If \code{num.clusters} and \code{num.groups} are not defined,
best clustering distribution is used to create the train set.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyBasedStrategy-plot"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyBasedStrategy-plot}{}}}
\subsection{Method \code{plot()}}{
The function is responsible for creating a plot to visualize
the clustering distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$plot(dir.path = NULL, file.name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{An optional argument to define the name of the directory
where the exported plot will be saved. If not defined, the file path will
be automatically assigned to the current working directory,
'\code{getwd()}'.}

\item{\code{file.name}}{A character to define the name of the PDF file where the
plot is exported.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyBasedStrategy-saveCSV"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyBasedStrategy-saveCSV}{}}}
\subsection{Method \code{saveCSV()}}{
The function is used to save the clustering distribution to
a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$saveCSV(
  dir.path = NULL,
  name = NULL,
  num.clusters = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir.path}}{The name of the directory to save the CSV file.}

\item{\code{name}}{Defines the name of the CSV file.}

\item{\code{num.clusters}}{An optional parameter to select the number of
clusters to be saved. If not defined, all cluster distributions will be
saved.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyBasedStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyBasedStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
