% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/clustering.strategies.DependencyBasedStrategyConfiguration.R
\name{DependencyBasedStrategyConfiguration}
\alias{DependencyBasedStrategyConfiguration}
\title{Custom Strategy Configuration handler for the DependencyBasedStrategy strategy.}
\description{
Define the default configuration parameters for the \link{DependencyBasedStrategy} strategy.
}
\seealso{
\code{\link{StrategyConfiguration}}, \code{\link{DependencyBasedStrategy}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[DDMCS:StrategyConfiguration]{DDMCS::StrategyConfiguration}} -> \code{DependencyBasedStrategyConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DependencyBasedStrategyConfiguration$new()}}
\item \href{#method-minNumClusters}{\code{DependencyBasedStrategyConfiguration$minNumClusters()}}
\item \href{#method-maxNumClusters}{\code{DependencyBasedStrategyConfiguration$maxNumClusters()}}
\item \href{#method-getBinaryCutoff}{\code{DependencyBasedStrategyConfiguration$getBinaryCutoff()}}
\item \href{#method-getRealCutoff}{\code{DependencyBasedStrategyConfiguration$getRealCutoff()}}
\item \href{#method-tiebreak}{\code{DependencyBasedStrategyConfiguration$tiebreak()}}
\item \href{#method-qualityOfCluster}{\code{DependencyBasedStrategyConfiguration$qualityOfCluster()}}
\item \href{#method-isImprovingClustering}{\code{DependencyBasedStrategyConfiguration$isImprovingClustering()}}
\item \href{#method-clone}{\code{DependencyBasedStrategyConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Empty function used to initalize the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \link{StrategyConfiguration} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-minNumClusters"></a>}}
\subsection{Method \code{minNumClusters()}}{
Function used to return the minimum number of clusters distributions used. By default the minimum is set in 2.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$minNumClusters(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Further arguments passed down to \code{minNumClusters} function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-maxNumClusters"></a>}}
\subsection{Method \code{maxNumClusters()}}{
The function is responsible of returning the maximum number of cluster distributions used.
By default the maximum number is set in 50.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$maxNumClusters(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Further arguments passed down to \code{maxNumClusters} function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBinaryCutoff"></a>}}
\subsection{Method \code{getBinaryCutoff()}}{
The function is used to define the interval to consider the dependency between binary features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$getBinaryCutoff()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Further arguments passed down to \code{maxNumClusters} function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRealCutoff"></a>}}
\subsection{Method \code{getRealCutoff()}}{
The function allows defining the cutoff to consider the dependency between real features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$getRealCutoff()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Further arguments passed down to \code{maxNumClusters} function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tiebreak"></a>}}
\subsection{Method \code{tiebreak()}}{
The function solves the ties between two (or more) features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$tiebreak(
  feature,
  clus.candidates,
  fea.dep.dist.clus,
  corpus,
  heuristic,
  class,
  class.name
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{A \link{character} containing the name of the feature}

\item{\code{clus.candidates}}{A single or \link{numeric} vector value to identify
the candidate groups to insert the feature.}

\item{\code{fea.dep.dist.clus}}{A \link{list} containing the groups chosen for the features.}

\item{\code{corpus}}{A \link{data.frame} containing the features of the initial data.}

\item{\code{heuristic}}{The heuristic used to compute the relevance of each feature.
Must inherit from \link{GenericHeuristic} abstract class.}

\item{\code{class}}{a \link{character} vector containing all the values of the target class.}

\item{\code{class.name}}{A \link{character} value representing the name of the target class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-qualityOfCluster"></a>}}
\subsection{Method \code{qualityOfCluster()}}{
The function determines the quality of a cluster.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$qualityOfCluster(clusters, metrics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusters}}{A \link{list} with the feature distribution of each cluster.}

\item{\code{metrics}}{A numeric \link{list} with the metrics associated to the cluster (dependency between all features and dependency between the features and the class).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isImprovingClustering"></a>}}
\subsection{Method \code{isImprovingClustering()}}{
The function indicates if clustering is getting better as the number of them increases.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$isImprovingClustering(clusters.deltha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusters.deltha}}{A \link{numeric} vector value with the quality values of the built clusters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyBasedStrategyConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
