% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.heuristics.GainRatioHeuristic.R
\name{GainRatioHeuristic}
\alias{GainRatioHeuristic}
\title{Feature-clustering based on GainRatio methodology.}
\description{
Performs the feature-clustering using entropy-based filters.
}
\seealso{
\code{\link{Dataset}}, \code{\link{gain.ratio}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[D2MCS:GenericHeuristic]{D2MCS::GenericHeuristic}} -> \code{GainRatioHeuristic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GainRatioHeuristic-new}{\code{GainRatioHeuristic$new()}}
\item \href{#method-GainRatioHeuristic-heuristic}{\code{GainRatioHeuristic$heuristic()}}
\item \href{#method-GainRatioHeuristic-clone}{\code{GainRatioHeuristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GainRatioHeuristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-GainRatioHeuristic-new}{}}}
\subsection{Method \code{new()}}{
Empty function used to initialize the object arguments in
runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GainRatioHeuristic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GainRatioHeuristic-heuristic"></a>}}
\if{latex}{\out{\hypertarget{method-GainRatioHeuristic-heuristic}{}}}
\subsection{Method \code{heuristic()}}{
The algorithms find weights of discrete attributes basing on
their correlation with continuous class attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GainRatioHeuristic$heuristic(col1, col2, column.names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{A \link{numeric} vector or matrix required to perform the
clustering operation.}

\item{\code{col2}}{A \link{numeric} vector or matrix to perform the clustering
operation.}

\item{\code{column.names}}{An optional \link{character} vector with the names of
both columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1 or \link{NA} if an error
occurs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GainRatioHeuristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GainRatioHeuristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GainRatioHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
