% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.utility.GenericModelFit.R
\name{GenericModelFit}
\alias{GenericModelFit}
\title{Abstract class for defining model fitting method.}
\description{
Template to create a \code{\link[recipes]{recipe}} or \code{\link{formula}} objects used in model training stage.
}
\seealso{
\code{\link{DefaultModelFit}}, \code{\link[caret]{train}}
}
\keyword{misc}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenericModelFit$new()}}
\item \href{#method-createFormula}{\code{GenericModelFit$createFormula()}}
\item \href{#method-createRecipe}{\code{GenericModelFit$createRecipe()}}
\item \href{#method-clone}{\code{GenericModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericModelFit$new(instances, class.name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instances}}{An \code{\link{data.frame}} containing the data used to train the M.L. models.}

\item{\code{class.name}}{A \code{\link{character}} of length 1 defining the name of the target class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createFormula"></a>}}
\subsection{Method \code{createFormula()}}{
The function is responsible of creating a \code{\link{formula}} for M.L. model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericModelFit$createFormula(simplify = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simplify}}{A \code{\link{logical}} argument defining whether the formula should be generated as simple as posible.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{formula}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createRecipe"></a>}}
\subsection{Method \code{createRecipe()}}{
The function is responsible of creating a \code{\link[recipes]{recipe}} for M.L. model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericModelFit$createRecipe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A object of class \code{\link[recipes]{recipe}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
