% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.plot.GenericPlot.R
\name{GenericPlot}
\alias{GenericPlot}
\title{Pseudo-abstract class for creating feature clustering plots.}
\description{
The \code{\link{GenericPlot}} implements a basic plot.
}
\seealso{
\code{\link{BinaryPlot}}
}
\keyword{color}
\keyword{device}
\keyword{hplot}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenericPlot-new}{\code{GenericPlot$new()}}
\item \href{#method-GenericPlot-plot}{\code{GenericPlot$plot()}}
\item \href{#method-GenericPlot-clone}{\code{GenericPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericPlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenericPlot-new}{}}}
\subsection{Method \code{new()}}{
Empty function used to initialize the object arguments in
runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericPlot$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericPlot-plot"></a>}}
\if{latex}{\out{\hypertarget{method-GenericPlot-plot}{}}}
\subsection{Method \code{plot()}}{
Implements a generic plot to visualize basic
feature-clustering data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericPlot$plot(summary)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary}}{A \link{data.frame} comprising the elements to be plotted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericPlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenericPlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
