% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/clustering.heuristics.InformationGainHeuristic.R
\name{InformationGainHeuristic}
\alias{InformationGainHeuristic}
\title{Feature-clustering based on InformationGain methodology.}
\description{
Performs the feature-clustering using entropy-based filters.
}
\seealso{
\code{\link{Dataset}}, \code{\link[FSelector]{information.gain}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[D2MCS:GenericHeuristic]{D2MCS::GenericHeuristic}} -> \code{InformationGainHeuristic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InformationGainHeuristic-new}{\code{InformationGainHeuristic$new()}}
\item \href{#method-InformationGainHeuristic-heuristic}{\code{InformationGainHeuristic$heuristic()}}
\item \href{#method-InformationGainHeuristic-clone}{\code{InformationGainHeuristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InformationGainHeuristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-InformationGainHeuristic-new}{}}}
\subsection{Method \code{new()}}{
Empty function used to initialize the object arguments in
runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationGainHeuristic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InformationGainHeuristic-heuristic"></a>}}
\if{latex}{\out{\hypertarget{method-InformationGainHeuristic-heuristic}{}}}
\subsection{Method \code{heuristic()}}{
The algorithm find weights of discrete attributes basing on
their correlation with continuous class attribute. Particularly
Information Gain uses \code{H(Class) + H(Attribute) - H(Class, Attribute)}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationGainHeuristic$heuristic(col1, col2, column.names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{A \link{numeric} vector or matrix required to perform the
clustering operation.}

\item{\code{col2}}{A \link{numeric} vector or matrix to perform the clustering
operation.}

\item{\code{column.names}}{an optional \link{character} vector with the names of
both columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1 or \link{NA} if an error
occurs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InformationGainHeuristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InformationGainHeuristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationGainHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
