% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/clustering.heuristics.InformationGainHeuristic.R
\docType{class}
\name{InformationGainHeuristic}
\alias{InformationGainHeuristic}
\title{Feature-clustering based on InformationGain methodology.}
\description{
Performs the feature-clustering using entropy-based filters.
}
\seealso{
\code{\link{Dataset}}, \code{\link[FSelector]{information.gain}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[DDMCS:GenericHeuristic]{DDMCS::GenericHeuristic}} -> \code{InformationGainHeuristic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InformationGainHeuristic$new()}}
\item \href{#method-heuristic}{\code{InformationGainHeuristic$heuristic()}}
\item \href{#method-clone}{\code{InformationGainHeuristic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a \link{InformationGainHeuristic} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationGainHeuristic$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{InformationGainHeuristic} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-heuristic"></a>}}
\subsection{Method \code{heuristic()}}{
The algorithm find weights of discrete attributes basing on
their correlation with continous class attribute. Particularly
Information Gain uses \code{H(Class) + H(Attribute) - H(Class, Attribute)}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationGainHeuristic$heuristic(col1, col2, column.names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{a numeric vector or matrix required to perform the
clustering operation.}

\item{\code{col2}}{a numeric vector or matrix to perform the clustering operation.}

\item{\code{column.names}}{an optional \link{character} vector with the names of
both columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \link{numeric} vector of length 1 or \link{NA} if an error occurs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationGainHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
