% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.heuristics.KendallHeuristic.R
\docType{class}
\name{KendallHeuristic}
\alias{KendallHeuristic}
\title{Feature-clustering based on Kendall Correlation Test.}
\description{
Performs the feature-clustering using Kendall correlation tests.
}
\details{
The method estimate the association between paired samples and
compute a test of the value being zero. They use different measures of association,
all in the range [-1, 1] with 0 indicating no association.
Method valid only for bi-class problems.
}
\seealso{
\code{\link{Dataset}}, \link{cor.test}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[DDMCS:GenericHeuristic]{DDMCS::GenericHeuristic}} -> \code{KendallHeuristic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KendallHeuristic$new()}}
\item \href{#method-heuristic}{\code{KendallHeuristic$heuristic()}}
\item \href{#method-clone}{\code{KendallHeuristic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a \link{KendallHeuristic} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KendallHeuristic$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{KendallHeuristic} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-heuristic"></a>}}
\subsection{Method \code{heuristic()}}{
Test for association between paired samples using Kendall's tau value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KendallHeuristic$heuristic(col1, col2, column.names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{a numeric vector or matrix required to perform the
clustering operation.}

\item{\code{col2}}{a numeric vector or matrix to perform the clustering operation.}

\item{\code{column.names}}{an optional \link{character} vector with the names of
both columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \link{numeric} vector of length 1 or \link{NA} if an error occurs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KendallHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
