% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.MCC.R
\name{MCC}
\alias{MCC}
\title{Computes the Matthews correlation coefficient.}
\description{
The Matthews correlation coefficient is used in machine learning
as a measure of the quality of binary (two-class) classifications. It takes
into account true and false positives and negatives and is generally regarded
as a balanced measure which can be used even if the classes are of very
different sizes. The MCC is in essence a correlation coefficient between the
observed and predicted binary classifications; it returns a value between -1
and +1.
}
\details{
\deqn{MCC = (TP × (TN - FP) × FN)/(\sqrt{(TP + FP) × (TP + FN) × (TN + FP) × (TN + FN)})}
}
\seealso{
\code{\link{MeasureFunction}}, \code{\link{ClassificationOutput}},
\code{\link{ConfMatrix}}
}
\keyword{classif}
\keyword{math}
\section{Super class}{
\code{\link[D2MCS:MeasureFunction]{D2MCS::MeasureFunction}} -> \code{MCC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MCC-new}{\code{MCC$new()}}
\item \href{#method-MCC-compute}{\code{MCC$compute()}}
\item \href{#method-MCC-clone}{\code{MCC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCC-new"></a>}}
\if{latex}{\out{\hypertarget{method-MCC-new}{}}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCC$new(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
used as basis to compute the \code{MCC} measure.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCC-compute"></a>}}
\if{latex}{\out{\hypertarget{method-MCC-compute}{}}}
\subsection{Method \code{compute()}}{
The function computes the \strong{MCC} achieved by the M.L.
model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCC$compute(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the \code{MCC}
measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is automatically invoke by the
\link{ClassificationOutput} object.
}

\subsection{Returns}{
A \link{numeric} vector of size 1 or \link{NULL} if an error
occurred.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MCC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
