% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.summaryFunction.NoProbability.R
\name{NoProbability}
\alias{NoProbability}
\title{Compute performance across resamples.}
\description{
Computes the performance across resamples when class
probabilities cannot be computed.
}
\seealso{
\code{\link{SummaryFunction}}
}
\keyword{misc}
\section{Super class}{
\code{\link[D2MCS:SummaryFunction]{D2MCS::SummaryFunction}} -> \code{NoProbability}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NoProbability-new}{\code{NoProbability$new()}}
\item \href{#method-NoProbability-execute}{\code{NoProbability$execute()}}
\item \href{#method-NoProbability-clone}{\code{NoProbability$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="D2MCS" data-topic="SummaryFunction" data-id="getMeasures"><a href='../../D2MCS/html/SummaryFunction.html#method-SummaryFunction-getMeasures'><code>D2MCS::SummaryFunction$getMeasures()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NoProbability-new"></a>}}
\if{latex}{\out{\hypertarget{method-NoProbability-new}{}}}
\subsection{Method \code{new()}}{
The function defined during runtime the usage of five
measures: 'Kappa', 'Accuracy', 'TCR_9', 'MCC' and 'PPV'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NoProbability-execute"></a>}}
\if{latex}{\out{\hypertarget{method-NoProbability-execute}{}}}
\subsection{Method \code{execute()}}{
The function computes the performance across resamples using
the previously defined measures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$execute(data, lev = NULL, model = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \link{data.frame} containing the data used to
compute the performance.}

\item{\code{lev}}{An optional value used to define the levels of the target
class.}

\item{\code{model}}{An optional value used to define the M.L. model used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of performance estimates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NoProbability-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NoProbability-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
