% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.summaryFunction.NoProbability.R
\name{NoProbability}
\alias{NoProbability}
\title{Compute performance across resamples.}
\description{
Computes the performance across resamples when class probabilities cannot be computed.
}
\seealso{
\code{\link{SummaryFunction}}
}
\keyword{misc}
\section{Super class}{
\code{\link[DDMCS:SummaryFunction]{DDMCS::SummaryFunction}} -> \code{NoProbability}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NoProbability$new()}}
\item \href{#method-execute}{\code{NoProbability$execute()}}
\item \href{#method-clone}{\code{NoProbability$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SummaryFunction" data-id="getMeasures">}\href{../../DDMCS/html/SummaryFunction.html#method-getMeasures}{\code{DDMCS::SummaryFunction$getMeasures()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
The function defined during runtime the usage of five measures: 'Kappa', 'Accuracy', 'TCR_9', 'MCC' and 'PPV'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link{NoProbability}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
The function computes the performance across resamples using the previously defined measures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$execute(data, lev = NULL, model = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{\link{data.frame}} containing the data used to compute the performance.}

\item{\code{lev}}{An optional value used to define the levels of the target class.}

\item{\code{model}}{An optional value used to define the M.L. model used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of performance estimates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoProbability$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
