% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.heuristics.OddsRatioHeuristic.R
\name{OddsRatioHeuristic}
\alias{OddsRatioHeuristic}
\title{Feature-clustering based on Odds Ratio measure.}
\description{
Performs the feature-clustering using Odds Ratio methodology.
Valid only for bi-class problems.
}
\seealso{
\code{\link{Dataset}}, \code{\link[questionr]{odds.ratio}}
}
\keyword{cluster}
\keyword{manip}
\section{Super class}{
\code{\link[D2MCS:GenericHeuristic]{D2MCS::GenericHeuristic}} -> \code{OddsRatioHeuristic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OddsRatioHeuristic-new}{\code{OddsRatioHeuristic$new()}}
\item \href{#method-OddsRatioHeuristic-heuristic}{\code{OddsRatioHeuristic$heuristic()}}
\item \href{#method-OddsRatioHeuristic-clone}{\code{OddsRatioHeuristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OddsRatioHeuristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-OddsRatioHeuristic-new}{}}}
\subsection{Method \code{new()}}{
Empty function used to initialize the object arguments in
runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OddsRatioHeuristic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OddsRatioHeuristic-heuristic"></a>}}
\if{latex}{\out{\hypertarget{method-OddsRatioHeuristic-heuristic}{}}}
\subsection{Method \code{heuristic()}}{
Calculates the Odds Ratio method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OddsRatioHeuristic$heuristic(col1, col2, column.names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col1}}{The object from whom odds ratio will be computed.}

\item{\code{col2}}{A second \link{factor} or \link{numeric} object.}

\item{\code{column.names}}{An optional \link{character} vector with the names of
both columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{numeric} vector of length 1 or \link{NA} if an error
occurs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OddsRatioHeuristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OddsRatioHeuristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OddsRatioHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
