% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.PPV.R
\name{PPV}
\alias{PPV}
\title{Computes the Positive Predictive Value.}
\description{
Positive Predictive Values are the proportions of positive
results in statistics and diagnostic tests that are true positive results.
}
\details{
\deqn{PPV = TP / (TP + FP)}
}
\seealso{
\code{\link{MeasureFunction}}, \code{\link{ClassificationOutput}},
\code{\link{ConfMatrix}}
}
\keyword{classif}
\keyword{math}
\section{Super class}{
\code{\link[D2MCS:MeasureFunction]{D2MCS::MeasureFunction}} -> \code{PPV}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PPV-new}{\code{PPV$new()}}
\item \href{#method-PPV-compute}{\code{PPV$compute()}}
\item \href{#method-PPV-clone}{\code{PPV$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPV-new"></a>}}
\if{latex}{\out{\hypertarget{method-PPV-new}{}}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPV$new(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the \strong{PPV}
measure.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPV-compute"></a>}}
\if{latex}{\out{\hypertarget{method-PPV-compute}{}}}
\subsection{Method \code{compute()}}{
The function computes the \strong{PPV} achieved by the M.L.
model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPV$compute(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the \strong{PPV}
measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is automatically invoke by the
\link{ClassificationOutput} object.
}

\subsection{Returns}{
A \link{numeric} vector of size 1 or \link{NULL} if an error
occurred.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PPV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
