% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.Precision.R
\name{Precision}
\alias{Precision}
\title{Computes the Precision Value.}
\description{
Precision is the fraction of relevant instances among the
retrieved instances
}
\details{
\deqn{precision = TP / (TP + FP)}
}
\seealso{
\code{\link{MeasureFunction}}, \code{\link{ClassificationOutput}},
\code{\link{ConfMatrix}}
}
\keyword{classif}
\keyword{math}
\section{Super class}{
\code{\link[D2MCS:MeasureFunction]{D2MCS::MeasureFunction}} -> \code{Precision}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Precision-new}{\code{Precision$new()}}
\item \href{#method-Precision-compute}{\code{Precision$compute()}}
\item \href{#method-Precision-clone}{\code{Precision$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Precision-new"></a>}}
\if{latex}{\out{\hypertarget{method-Precision-new}{}}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Precision$new(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the measure.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Precision-compute"></a>}}
\if{latex}{\out{\hypertarget{method-Precision-compute}{}}}
\subsection{Method \code{compute()}}{
The function computes the \strong{Precision} achieved by the
M.L. model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Precision$compute(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the
\strong{Precision} measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is automatically invoke by the
\link{ClassificationOutput} object.
}

\subsection{Returns}{
A \link{numeric} vector of size 1 or \link{NULL} if an error
occurred.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Precision-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Precision-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Precision$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
