% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.ProbAverageVoting.R
\docType{class}
\name{ProbAverageVoting}
\alias{ProbAverageVoting}
\title{<<tittle>>}
\description{
ProbAverageVoting
}
\details{
<<details>
}
\seealso{
\code{\link{SimpleVoting}}
}
\section{Super class}{
\code{\link[DDMCS:SimpleVoting]{DDMCS::SimpleVoting}} -> \code{ProbAverageVoting}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProbAverageVoting$new()}}
\item \href{#method-getMajorityClass}{\code{ProbAverageVoting$getMajorityClass()}}
\item \href{#method-getClassTie}{\code{ProbAverageVoting$getClassTie()}}
\item \href{#method-execute}{\code{ProbAverageVoting$execute()}}
\item \href{#method-clone}{\code{ProbAverageVoting$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SimpleVoting" data-id="getCutoff">}\href{../../DDMCS/html/SimpleVoting.html#method-getCutoff}{\code{DDMCS::SimpleVoting$getCutoff()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="SimpleVoting" data-id="getFinalPred">}\href{../../DDMCS/html/SimpleVoting.html#method-getFinalPred}{\code{DDMCS::SimpleVoting$getFinalPred()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageVoting$new(cutoff = 0.5, class.tie = NULL, majority.class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cutoff}}{<<description>>}

\item{\code{class.tie}}{<<description>>}

\item{\code{majority.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMajorityClass"></a>}}
\subsection{Method \code{getMajorityClass()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageVoting$getMajorityClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassTie"></a>}}
\subsection{Method \code{getClassTie()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageVoting$getClassTie()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageVoting$execute(predictions, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{<<description>>}

\item{\code{verbose}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbAverageVoting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
