% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.ProbBasedMethodology.R
\docType{class}
\name{ProbBasedMethodology}
\alias{ProbBasedMethodology}
\title{<<tittle>>}
\description{
ProbBasedMethodology
}
\details{
<<details>
}
\seealso{
\code{\link{Methodology}}
}
\section{Super class}{
\code{\link[DDMCS:Methodology]{DDMCS::Methodology}} -> \code{ProbBasedMethodology}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProbBasedMethodology$new()}}
\item \href{#method-compute}{\code{ProbBasedMethodology$compute()}}
\item \href{#method-clone}{\code{ProbBasedMethodology$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="DDMCS" data-topic="Methodology" data-id="getRequiredMetrics">}\href{../../DDMCS/html/Methodology.html#method-getRequiredMetrics}{\code{DDMCS::Methodology$getRequiredMetrics()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbBasedMethodology$new(required.metrics = c("MCC", "PPV"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{required.metrics}}{<<description>>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\subsection{Method \code{compute()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbBasedMethodology$compute(
  raw.pred,
  prob.pred,
  positive.class,
  negative.class
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw.pred}}{<<description>>}

\item{\code{prob.pred}}{<<description>>}

\item{\code{positive.class}}{<<description>>}

\item{\code{negative.class}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbBasedMethodology$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
