% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.Sensitivity.R
\name{Sensitivity}
\alias{Sensitivity}
\title{Computes the Sensitivity Value.}
\description{
Sensitivity is a measure of the proportion of actual positive
cases that got predicted as positive (or true positive).
}
\details{
\deqn{Sensitivity = TP / (TP + FN)}
}
\seealso{
\code{\link{MeasureFunction}}, \code{\link{ClassificationOutput}},
\code{\link{ConfMatrix}}
}
\keyword{classif}
\keyword{math}
\section{Super class}{
\code{\link[D2MCS:MeasureFunction]{D2MCS::MeasureFunction}} -> \code{Sensitivity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sensitivity-new}{\code{Sensitivity$new()}}
\item \href{#method-Sensitivity-compute}{\code{Sensitivity$compute()}}
\item \href{#method-Sensitivity-clone}{\code{Sensitivity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sensitivity-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sensitivity-new}{}}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sensitivity$new(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the
\code{Sensitivity} measure.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sensitivity-compute"></a>}}
\if{latex}{\out{\hypertarget{method-Sensitivity-compute}{}}}
\subsection{Method \code{compute()}}{
The function computes the \strong{Sensitivity} achieved by the
M.L. model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sensitivity$compute(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the
\strong{Sensitivity} measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is automatically invoke by the
\link{ClassificationOutput} object.
}

\subsection{Returns}{
A \link{numeric} vector of size 1 or \link{NULL} if
an error occurred.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sensitivity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sensitivity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sensitivity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
