% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.Specificity.R
\name{Specificity}
\alias{Specificity}
\title{Computes the Specificity Value.}
\description{
Specificity is defined as the proportion of actual negatives,
which got predicted as the negative (or true negative). This implies that
there will be another proportion of actual negative, which got predicted as
positive and could be termed as false positives.
}
\details{
\deqn{Specificity = True Negative / (True Negative + False Positive)}
}
\seealso{
\code{\link{MeasureFunction}}, \code{\link{ClassificationOutput}},
\code{\link{ConfMatrix}}
}
\keyword{classif}
\keyword{math}
\section{Super class}{
\code{\link[D2MCS:MeasureFunction]{D2MCS::MeasureFunction}} -> \code{Specificity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Specificity-new}{\code{Specificity$new()}}
\item \href{#method-Specificity-compute}{\code{Specificity$compute()}}
\item \href{#method-Specificity-clone}{\code{Specificity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specificity-new"></a>}}
\if{latex}{\out{\hypertarget{method-Specificity-new}{}}}
\subsection{Method \code{new()}}{
Method for initializing the object arguments during runtime.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specificity$new(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the measure.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specificity-compute"></a>}}
\if{latex}{\out{\hypertarget{method-Specificity-compute}{}}}
\subsection{Method \code{compute()}}{
The function computes the \strong{Specificity} achieved by the
M.L. model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specificity$compute(performance.output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{performance.output}}{An optional \code{\link{ConfMatrix}} parameter
to define the type of object used as basis to compute the
\strong{Specificity} measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is automatically invoke by the
\code{\link{ClassificationOutput}} object.
}

\subsection{Returns}{
A \link{numeric} vector of size 1 or \link{NULL} if an error
occurred.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specificity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Specificity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specificity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
