% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.Trainset.R
\docType{class}
\name{Trainset}
\alias{Trainset}
\title{Trainning set.}
\description{
The \code{Trainset} is used to perform training operations over M.L. models.
A target class should be defined to guarantee a full compatibility with supervised models.
}
\details{
Use \link{Dataset} object to ensure the creation of a valid \code{\link{Trainset}} object.
}
\seealso{
\link{Dataset}, \link{DatasetLoader}, \link{Subset}, \link{GenericClusteringStrategy}
}
\keyword{attribute}
\keyword{datasets}
\keyword{manip}
\keyword{programming}
\keyword{utilities}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Trainset$new()}}
\item \href{#method-getPositiveClass}{\code{Trainset$getPositiveClass()}}
\item \href{#method-getClassName}{\code{Trainset$getClassName()}}
\item \href{#method-getClassValues}{\code{Trainset$getClassValues()}}
\item \href{#method-getColumnNames}{\code{Trainset$getColumnNames()}}
\item \href{#method-getFeatureValues}{\code{Trainset$getFeatureValues()}}
\item \href{#method-getInstances}{\code{Trainset$getInstances()}}
\item \href{#method-getNumClusters}{\code{Trainset$getNumClusters()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a \link{Trainset} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainset$new(cluster.dist, class.name, class.values, positive.class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cluster.dist}}{the type of cluster distribution used as basis
to build the \link{Trainset}. See \link{GenericClusteringStrategy} for more information.}

\item{\code{class.name}}{used to specify the name of the column containing the target class.}

\item{\code{class.values}}{specifies all the posible values of the target class.}

\item{\code{positive.class}}{a \code{\link{character}} with the value of the positive class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a fully operative \link{Trainset} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPositiveClass"></a>}}
\subsection{Method \code{getPositiveClass()}}{
the function is used to obtain the value of the positive class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainset$getPositiveClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{numeric} value with the positive class value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassName"></a>}}
\subsection{Method \code{getClassName()}}{
the function is used to return the name of the target class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainset$getClassName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{character} vector with length 1.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClassValues"></a>}}
\subsection{Method \code{getClassValues()}}{
the function is used to compute all the posible target class values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainset$getClassValues()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{factor} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getColumnNames"></a>}}
\subsection{Method \code{getColumnNames()}}{
the function returns the name of the columns comprising
an specific cluster distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainset$getColumnNames(num.cluster)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.cluster}}{a \link{numeric} value used to specify the cluster number
of the cluster distribution used when creating the \link{Trainset}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \link{character} vector with all column names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeatureValues"></a>}}
\subsection{Method \code{getFeatureValues()}}{
the function returns the values of the columns comprising
an specific cluster distribution. Target class is omitted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainset$getFeatureValues(num.cluster)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.cluster}}{a \link{numeric} value used to specify the cluster number
of the cluster distribution used when creating the \link{Trainset}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \link{data.frame} with the values of the features comprising
the selected cluster distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInstances"></a>}}
\subsection{Method \code{getInstances()}}{
the function returns the values of the columns comprising
an specific cluster distribution. Target class is included as the last column.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainset$getInstances(num.cluster)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num.cluster}}{a \link{numeric} value used to specify the cluster number
of the cluster distribution used when creating the \link{Trainset}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \link{data.frame} with the values of the features comprising the
selected cluster distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNumClusters"></a>}}
\subsection{Method \code{getNumClusters()}}{
the function obtains the number of groups (clusters) that forms
the cluster distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainset$getNumClusters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link{numeric} vector of size 1.
}
}
}
