% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.VotingStrategy.R
\docType{class}
\name{VotingStrategy}
\alias{VotingStrategy}
\title{<<tittle>>}
\description{
VotingStrategy
}
\details{
<<details>
}
\seealso{
\code{\link{DDMCS}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VotingStrategy$new()}}
\item \href{#method-getVotingSchemes}{\code{VotingStrategy$getVotingSchemes()}}
\item \href{#method-getMetrics}{\code{VotingStrategy$getMetrics()}}
\item \href{#method-execute}{\code{VotingStrategy$execute()}}
\item \href{#method-getName}{\code{VotingStrategy$getName()}}
\item \href{#method-clone}{\code{VotingStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VotingStrategy$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVotingSchemes"></a>}}
\subsection{Method \code{getVotingSchemes()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VotingStrategy$getVotingSchemes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetrics"></a>}}
\subsection{Method \code{getMetrics()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VotingStrategy$getMetrics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VotingStrategy$execute(predictions, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{<<description>>}

\item{\code{...}}{<<description>>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\subsection{Method \code{getName()}}{
<<description>>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VotingStrategy$getName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<<description>>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VotingStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
